% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_function_template}
\alias{chk_function_template}
\title{Check Function Template}
\usage{
chk_function_template(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
\code{NULL}, invisibly. Called for the side effect of throwing an error
if the condition is not met.
}
\description{
A chk function template.
}
\details{
Checks if character using \verb{[vld_function_template](x)}.
}
\examples{
chk_function_template("1")
try(chk_function_template(1))
}
