% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-index.R
\name{chk_index}
\alias{chk_index}
\alias{vld_index}
\title{Check Index}
\usage{
chk_index(x, x_name = NULL)

vld_index(x)
}
\arguments{
\item{x}{An object.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if an object is a vector of one or more positive integer values.
}
\section{Functions}{
\itemize{
\item \code{vld_index}: Validate Index
}}

\examples{
x <- c(2L, 1L)
chk_index(x)
y <- c(2L, -1L)
try(chk_index(y))
vld_index(c(-1))
vld_index(c(3L, 1L))
}
