% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-natomic.R
\name{chk_natomic}
\alias{chk_natomic}
\alias{vld_natomic}
\title{Check Numeric Atomic Object}
\usage{
chk_natomic(x, x_name = NULL)

vld_natomic(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if numeric atomic object using
\code{is.numeric(x) && is.atomic(x)}.
}
\section{Functions}{
\itemize{
\item \code{vld_natomic}: Validate Numeric Atomic Object
}}

\examples{

# chk_natomic
chk_natomic(1)
chk_natomic(matrix(1L))
try(chk_natomic(TRUE))

#' vld_natomic
vld_natomic(1)
vld_natomic(matrix(1L))
try(vld_natomic(TRUE))
}
