% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_gamma}
\alias{dev_gamma}
\title{Gamma Deviances}
\usage{
dev_gamma(x, shape = 1, rate = 1, res = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{shape}{A non-negative numeric vector of shape.}

\item{rate}{A non-negative numeric vector of rate.}

\item{res}{A flag specifying whether to return the deviance residual as
opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Gamma Deviances
}
\examples{
dev_gamma(c(0, 1, 2), 1, 2)
}
\seealso{
Other dev_dist:
\code{\link[=dev_bern]{dev_bern()}},
\code{\link[=dev_beta_binom]{dev_beta_binom()}},
\code{\link[=dev_binom]{dev_binom()}},
\code{\link[=dev_gamma_pois]{dev_gamma_pois()}},
\code{\link[=dev_lnorm]{dev_lnorm()}},
\code{\link[=dev_multinom]{dev_multinom()}},
\code{\link[=dev_neg_binom]{dev_neg_binom()}},
\code{\link[=dev_norm]{dev_norm()}},
\code{\link[=dev_pois]{dev_pois()}},
\code{\link[=dev_pois_zi]{dev_pois_zi()}},
\code{\link[=dev_skewlnorm]{dev_skewlnorm()}},
\code{\link[=dev_skewnorm]{dev_skewnorm()}},
\code{\link[=dev_student]{dev_student()}}
}
\concept{dev_dist}
