% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_gamma_pois_zi}
\alias{dev_gamma_pois_zi}
\title{Zero-Inflated Gamma-Poisson Deviances}
\usage{
dev_gamma_pois_zi(x, lambda = 1, theta = 0, prob = 0, res = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the shape and rate.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Zero-Inflated Gamma-Poisson Deviances
}
\examples{
dev_gamma_pois_zi(c(1,3.5,4), 3, 2)
}
\concept{dev_dist # make live when complete}
