% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_gamma_pois_zt}
\alias{dev_gamma_pois_zt}
\title{Zero-Truncated Gamma-Poisson Deviances}
\usage{
dev_gamma_pois_zt(x, lambda = 1, theta = 0, res = FALSE)
}
\arguments{
\item{x}{A whole numeric vector of values greater than or equal to 1.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}

\item{res}{A flag specifying whether to return the deviance residual as
opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Saturated log-likelihood is found numerically per observation since the
truncated likelihood has no closed-form maximum.
}
\examples{
dev_gamma_pois_zt(c(1, 3, 4), 3, 2)
}
\seealso{
Other dev_dist # make live when complete: 
\code{\link{dev_gamma_pois_zi}()}
}
\concept{dev_dist # make live when complete}
