% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_lnorm_hurdle}
\alias{dev_lnorm_hurdle}
\title{Log-Normal Hurdle Deviances}
\usage{
dev_lnorm_hurdle(x, meanlog = 0, sdlog = 1, prob = 0, res = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the log scale.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Log-Normal Hurdle Deviances
}
\examples{
dev_lnorm(exp(-2:2))
}
\seealso{
Other dev_dist: 
\code{\link{dev_bern}()},
\code{\link{dev_beta_binom}()},
\code{\link{dev_binom}()},
\code{\link{dev_gamma_pois}()},
\code{\link{dev_gamma}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_neg_binom}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois_zi}()},
\code{\link{dev_pois}()},
\code{\link{dev_student}()}
}
\concept{dev_dist}
