% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_neg_binom}
\alias{dev_neg_binom}
\title{Negative Binomial Deviances}
\usage{
dev_neg_binom(x, lambda = 1, theta = 0, res = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}

\item{res}{A flag specifying whether to return the deviance residual as
opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Negative Binomial Deviances
}
\examples{
dev_neg_binom(c(1, 2, 5), 2, 3)
}
\seealso{
Other dev_dist:
\code{\link[=dev_bern]{dev_bern()}},
\code{\link[=dev_beta_binom]{dev_beta_binom()}},
\code{\link[=dev_binom]{dev_binom()}},
\code{\link[=dev_gamma]{dev_gamma()}},
\code{\link[=dev_gamma_pois]{dev_gamma_pois()}},
\code{\link[=dev_lnorm]{dev_lnorm()}},
\code{\link[=dev_multinom]{dev_multinom()}},
\code{\link[=dev_norm]{dev_norm()}},
\code{\link[=dev_pois]{dev_pois()}},
\code{\link[=dev_pois_zi]{dev_pois_zi()}},
\code{\link[=dev_skewlnorm]{dev_skewlnorm()}},
\code{\link[=dev_skewnorm]{dev_skewnorm()}},
\code{\link[=dev_student]{dev_student()}}
}
\concept{dev_dist}
