% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_skewlnorm}
\alias{dev_skewlnorm}
\title{Skew-Lognormal Deviances}
\usage{
dev_skewlnorm(
  x,
  locationlog = 0,
  scalelog = 1,
  shapelog = 0,
  res = FALSE,
  ...,
  meanlog,
  sdlog,
  shape
)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{locationlog}{A numeric vector of location parameters of \code{log(x)}.}

\item{scalelog}{A non-negative numeric vector of scale parameters of \code{log(x)}.}

\item{shapelog}{A numeric vector of shape parameters of \code{log(x)}.
Negative values result in leftward skew, while positive values result in rightward skew.}

\item{res}{A flag specifying whether to return the deviance residual as
opposed to the deviance.}

\item{...}{Other arguments passed to methods.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the
log scale.}

\item{shape}{A non-negative numeric vector of shape.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Skew-Lognormal Deviances
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
dev_skewlnorm(exp(-2:2))
dev_skewlnorm(exp(-2:2), 0, 1, 5)
dev_skewlnorm(exp(-2:2), 0, 1, 5, res = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other dev_dist:
\code{\link[=dev_bern]{dev_bern()}},
\code{\link[=dev_beta_binom]{dev_beta_binom()}},
\code{\link[=dev_binom]{dev_binom()}},
\code{\link[=dev_gamma]{dev_gamma()}},
\code{\link[=dev_gamma_pois]{dev_gamma_pois()}},
\code{\link[=dev_lnorm]{dev_lnorm()}},
\code{\link[=dev_multinom]{dev_multinom()}},
\code{\link[=dev_neg_binom]{dev_neg_binom()}},
\code{\link[=dev_norm]{dev_norm()}},
\code{\link[=dev_pois]{dev_pois()}},
\code{\link[=dev_pois_zi]{dev_pois_zi()}},
\code{\link[=dev_skewnorm]{dev_skewnorm()}},
\code{\link[=dev_student]{dev_student()}}
}
\concept{dev_dist}
