% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_student}
\alias{dev_student}
\title{Student's t Deviances}
\usage{
dev_student(x, mean = 0, sd = 1, theta = 0, res = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}

\item{res}{A flag specifying whether to return the deviance residual as
opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Student's t Deviances
}
\examples{
dev_student(c(1, 3.5, 4), 3)
}
\seealso{
Other dev_dist:
\code{\link[=dev_bern]{dev_bern()}},
\code{\link[=dev_beta_binom]{dev_beta_binom()}},
\code{\link[=dev_binom]{dev_binom()}},
\code{\link[=dev_gamma]{dev_gamma()}},
\code{\link[=dev_gamma_pois]{dev_gamma_pois()}},
\code{\link[=dev_lnorm]{dev_lnorm()}},
\code{\link[=dev_multinom]{dev_multinom()}},
\code{\link[=dev_neg_binom]{dev_neg_binom()}},
\code{\link[=dev_norm]{dev_norm()}},
\code{\link[=dev_pois]{dev_pois()}},
\code{\link[=dev_pois_zi]{dev_pois_zi()}},
\code{\link[=dev_skewlnorm]{dev_skewlnorm()}},
\code{\link[=dev_skewnorm]{dev_skewnorm()}}
}
\concept{dev_dist}
