% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{devbern}
\alias{devbern}
\title{Bernoulli Deviance}
\usage{
devbern(x, prob, residual = FALSE)
}
\arguments{
\item{x}{A logical vector.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of TRUE.}

\item{residual}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Bernoulli Deviance
}
\examples{
devbern(c(TRUE, FALSE), 0.7, residual = TRUE)
}
\seealso{
Other devdist: 
\code{\link{devbinom}()},
\code{\link{devlnorm}()},
\code{\link{devnorm}()},
\code{\link{devpois}()}
}
\concept{devdist}
