% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-residuals.R
\name{devbinom}
\alias{devbinom}
\title{Binomial Deviance Residual}
\usage{
devbinom(x, size, prob)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{size}{A non-negative whole numeric vector of the number of trials.}

\item{prob}{A numeric vector of values between 0 and 1 o the probability of success of each trial.}
}
\value{
An numeric vector of the corresponding deviance residuals.
}
\description{
Binomial Deviance Residual
}
\examples{
devbinom(c(0, 1, 2), 2, 0.3)
}
\seealso{
Other devdist: 
\code{\link{devbern}()},
\code{\link{devlnorm}()},
\code{\link{devnorm}()},
\code{\link{devpois}()}
}
\concept{devdist}
