% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{devlnorm}
\alias{devlnorm}
\title{Log-Normal Deviance}
\usage{
devlnorm(x, meanlog = 0, sdlog = 1, residual = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the log scale.}

\item{residual}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Log-Normal Deviance
}
\examples{
devlnorm(exp(-2:2), residual = TRUE)
}
\seealso{
Other devdist: 
\code{\link{devbern}()},
\code{\link{devbinom}()},
\code{\link{devnorm}()},
\code{\link{devpois}()}
}
\concept{devdist}
