% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{devnorm}
\alias{devnorm}
\title{Normal Deviance}
\usage{
devnorm(x, mean = 0, sd = 1, residual = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{residual}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Normal Deviance
}
\examples{
devnorm(c(-2:2), residual = TRUE)
}
\seealso{
Other devdist: 
\code{\link{devbern}()},
\code{\link{devbinom}()},
\code{\link{devlnorm}()},
\code{\link{devpois}()}
}
\concept{devdist}
