% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-residuals.R
\name{devpois}
\alias{devpois}
\title{Poisson Deviance Residual}
\usage{
devpois(x, lambda)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}
}
\value{
An numeric vector of the corresponding deviance residuals.
}
\description{
Poisson Deviance Residual
}
\examples{
devpois(c(1,3.5,4), 3)
}
