% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.data.frame}
\alias{dims.data.frame}
\title{Dimensions}
\usage{
\method{dims}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
}
\details{
Unlike \code{base::dim()}, dims returns the length of an atomic vector.
}
\examples{
dims(data.frame())
dims(data.frame(x = 1:3))
}
\seealso{
\code{\link[base:dim]{base::dim()}}

Other {dimensions}: 
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims_terms}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims_terms}()},
\code{\link[universals]{pdims}()},
\code{\link[universals]{tindex}()}
}
