% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.default}
\alias{dims.default}
\title{Dimensions of a default object}
\usage{
\method{dims}{default}(x, ...)
}
\arguments{
\item{x}{An object. See the individual methods for specifics.}

\item{...}{Other arguments passed to methods.}
}
\description{
Dimensions of a default object
}
\examples{
dims(character(0))
dims(1:3)
}
\seealso{
\code{\link[universals]{dims}()}
}
