% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\title{Direction of a distribution}
\usage{
direction(x, estimate = median, threshold = 0, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of MCMC values or any other numeric vector of samples.}

\item{estimate}{A function for estimating the center of the distribution.
Defaults to \code{\link[=median]{median()}}, but can also be \verb{[mean()}] or any custom function
that returns a numeric vector of length 1.
\code{NA} values are dropped before calling the function.}

\item{threshold}{A number of the threshold value.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A character vector of length one indicating if at least half of the
observations are above the threshold (\code{"right"}) or not (\code{"left"}).
}
\description{
The direction of a distribution is the side (left/right) that the
distribution's center falls on, relative to a threshold. The center can be
calculated using a user-specified function, including the \verb{[median]} (default),
\verb{[mean]}, geometric mean, mode, or any other custom function. By convention,
values below the threshold fall to the left, while values above the threshold
fall to the right. Center estimates equal to the threshold are assumed to
fall to the right.
}
\examples{
direction(c(1, 2, 3))
direction(c(-1))
direction(c(0, 0, 0))
direction(c(-100, 1, 1))
direction(c(-100, 1, 1), mean)
direction(c(100, 0.01, 0.01), function(.x) exp(mean(log(.x))))
}
\seealso{
Other summary: 
\code{\link{directional_information}()},
\code{\link{kurtosis}()},
\code{\link{lower}()},
\code{\link{probability_direction}()},
\code{\link{pvalue}()},
\code{\link{pzeros}()},
\code{\link{skewness}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{xtr_median}()},
\code{\link{xtr_sd}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
