% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional-information.R
\name{directional_information}
\alias{directional_information}
\title{Directional information}
\usage{
directional_information(
  x,
  side = "median",
  threshold = 0,
  threshold_split = "proportional",
  na_rm = FALSE
)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}

\item{side}{A string indicating whether to calculate
the directional information relative to the left side (\code{"left"}; \code{x < threshold}),
or the right side (\code{"right"}; \code{x > threshold}). Positive information suggests
greater evidence for the specified side.
Defaults to \code{"median"}, which uses the side of the median of \code{x} via
\code{\link[=direction]{direction()}}.}

\item{threshold}{A number of the threshold value.}

\item{threshold_split}{A string indicating how to deal
with threshold values:
\itemize{
\item \code{"left"} to include them on the left side,
\item \code{"right"} to include them on the right side,
\item \code{"equal"} to split them equally between the left and side,
\item \code{"proportional"} (default) to split them between the left and right sides
proportionally to the values of \code{x} on the left and right sides,
\item \code{"exclude"} to drop the values of \code{x} equal to \code{threshold}
(identical to using \code{"proportional"}).
}}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number indicating the directional information.
}
\description{
Quantifies the information about direction in a posterior distribution based on the directional probability.
This function calculates such information using the difference in the probability of direction (see \code{\link[=probability_direction]{probability_direction()}}), after converting each probability to bits (also see \code{\link[=svalue]{svalue()}}.
}
\examples{

directional_information(0)
directional_information(1) # one coin flip of information
directional_information(c(1, 1)) # two coin flips
directional_information(c(1, 1, -1)) # x[2] and x[3] cancel out
directional_information(c(1, 1, -1, -1)) # both sides cancel out
directional_information(rnorm(1e3, mean = 0))
directional_information(rnorm(1e3, mean = 1))
directional_information(rnorm(1e3, mean = 10)) # all coin flips are positive
directional_information(rnorm(1e3, mean = -10)) # all coin flips are negative
directional_information(rnorm(1e3, mean = 1e3)) # only quantiles matter
directional_information(rnorm(1e6, mean = 1e3)) # more `x` implies more info
}
\references{
Kery, M., and Schaub, M. 2011.
Bayesian population analysis using WinBUGS: a hierarchical perspective.
Academic Press, Boston. Available from \url{https://www.vogelwarte.ch/en/research/population-biology/book-bpa/}.
}
\seealso{
Other summary:
\code{\link[=direction]{direction()}},
\code{\link[=kurtosis]{kurtosis()}},
\code{\link[=lower]{lower()}},
\code{\link[=probability_direction]{probability_direction()}},
\code{\link[=pvalue]{pvalue()}},
\code{\link[=pzeros]{pzeros()}},
\code{\link[=skewness]{skewness()}},
\code{\link[=svalue]{svalue()}},
\code{\link[=upper]{upper()}},
\code{\link[=variance]{variance()}},
\code{\link[=xtr_mean]{xtr_mean()}},
\code{\link[=xtr_median]{xtr_median()}},
\code{\link[=xtr_sd]{xtr_sd()}},
\code{\link[=zeros]{zeros()}},
\code{\link[=zscore]{zscore()}}
}
\concept{summary}
