% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{exp10}
\alias{exp10}
\title{Exponential transformation of 10^x}
\usage{
exp10(x)
}
\arguments{
\item{x}{An numeric atomic object.}
}
\value{
A numeric atomic object.
}
\description{
Inverse of log10(x)
}
\details{
Returns the value of 10^x
}
\examples{
x <- 5
exp10(x) # 1e+05
10^x # 1e+05
}
\seealso{
Other translations: 
\code{\link{exp2}()},
\code{\link{fabs}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
