% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabs.R
\name{fabs}
\alias{fabs}
\title{Absolute}
\usage{
fabs(x)
}
\arguments{
\item{x}{An existing R object.}
}
\value{
A numeric vector of the corresponding absolute values.
}
\description{
Computes the absolute value of \code{x}.
Used in TMB as replacement for \code{abs()} which is seemingly ambiguous.
}
\details{
A wrapper on \code{\link{abs}}\verb{()}.
}
\examples{
fabs(c(0, -1, 2))
}
\seealso{
Other translations:
\code{\link[=exp10]{exp10()}},
\code{\link[=exp2]{exp2()}},
\code{\link[=ilog]{ilog()}},
\code{\link[=ilog10]{ilog10()}},
\code{\link[=ilog2]{ilog2()}},
\code{\link[=ilogit]{ilogit()}},
\code{\link[=inv_logit]{inv_logit()}},
\code{\link[=invlogit]{invlogit()}},
\code{\link[=log10<-]{log10<-()}},
\code{\link[=log2<-]{log2<-()}},
\code{\link[=log<-]{log<-()}},
\code{\link[=logit]{logit()}},
\code{\link[=logit<-]{logit<-()}},
\code{\link[=phi]{phi()}},
\code{\link[=pow]{pow()}},
\code{\link[=step]{step()}}
}
\concept{translations}
