% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na.character}
\alias{fill_na.character}
\title{Fill Missing Values}
\usage{
\method{fill_na}{character}(x, value = "0", ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace missing values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills an object's missing values while preserving the object's class.
}
\examples{
fill_na(c("text", NA))
fill_na(matrix(c("text", NA)), value = Inf)
}
\seealso{
Other {MCMC manipulations}: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{estimates}()},
\code{\link[universals]{fill_all}()},
\code{\link[universals]{split_chains}()}
}
