% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na.logical}
\alias{fill_na.logical}
\title{Fill Missing Values}
\usage{
\method{fill_na}{logical}(x, value = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace missing values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills an object's missing values.
}
\seealso{
Other {MCMC manipulations}: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{complete_terms}()},
\code{\link[universals]{fill_all}()},
\code{\link[universals]{split_chains}()}
}
