% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na.numeric}
\alias{fill_na.numeric}
\title{Fill Missing Values}
\usage{
\method{fill_na}{numeric}(x, value = 0, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace missing values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills an object's missing values while preserving the object's class.
}
\examples{
fill_na(c(1, NA), Inf)
}
