% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na}
\alias{fill_na}
\title{Fill Missing Values}
\usage{
fill_na(x, value, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace missing values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills an object's missing values while preserving the object's class.
}
