% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog2}
\alias{ilog2}
\title{Inverse Log Base 2 Transformation}
\usage{
ilog2(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A numeric atomic object.
}
\description{
Inverse log transforms a numeric atomic object with base 2.
}
\details{
A wrapper on \code{\link{exp2}}\code{(value)}.
}
\examples{
x <- c(2, 4.5)
ilog2(x)
}
\seealso{
Other translations:
\code{\link[=exp10]{exp10()}},
\code{\link[=exp2]{exp2()}},
\code{\link[=fabs]{fabs()}},
\code{\link[=ilog]{ilog()}},
\code{\link[=ilog10]{ilog10()}},
\code{\link[=ilogit]{ilogit()}},
\code{\link[=inv_logit]{inv_logit()}},
\code{\link[=invlogit]{invlogit()}},
\code{\link[=log10<-]{log10<-()}},
\code{\link[=log2<-]{log2<-()}},
\code{\link[=log<-]{log<-()}},
\code{\link[=logit]{logit()}},
\code{\link[=logit<-]{logit<-()}},
\code{\link[=phi]{phi()}},
\code{\link[=pow]{pow()}},
\code{\link[=step]{step()}}
}
\concept{translations}
