% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilogit.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logistic Transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A numeric atomic object.}
}
\value{
A numeric atomic object.
}
\description{
Inverse logistically transforms a numeric atomic object.
}
\details{
A wrapper on \code{\link[stats:plogis]{stats::plogis()}}.
}
\examples{
invlogit(c(-1, 0, 5))
}
\seealso{
Other translations:
\code{\link[=exp10]{exp10()}},
\code{\link[=exp2]{exp2()}},
\code{\link[=fabs]{fabs()}},
\code{\link[=ilog]{ilog()}},
\code{\link[=ilog10]{ilog10()}},
\code{\link[=ilog2]{ilog2()}},
\code{\link[=ilogit]{ilogit()}},
\code{\link[=inv_logit]{inv_logit()}},
\code{\link[=log10<-]{log10<-()}},
\code{\link[=log2<-]{log2<-()}},
\code{\link[=log<-]{log<-()}},
\code{\link[=logit]{logit()}},
\code{\link[=logit<-]{logit<-()}},
\code{\link[=phi]{phi()}},
\code{\link[=pow]{pow()}},
\code{\link[=step]{step()}}
}
\concept{translations}
