% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log2<-}
\alias{log2<-}
\title{Log Base 2 Transformation}
\usage{
log2(x) <- value
}
\arguments{
\item{x}{An object.}

\item{value}{A numeric atomic object.}
}
\value{
Called for the side effect of updating \code{x}.
}
\description{
Replaces a object with the base 2 exponent of value.
}
\details{
A wrapper on \code{\link{exp2}}\code{(value)}.
}
\examples{
x <- NULL
log2(x) <- c(0.5, 5)
x
}
\seealso{
Other translations:
\code{\link[=exp10]{exp10()}},
\code{\link[=exp2]{exp2()}},
\code{\link[=fabs]{fabs()}},
\code{\link[=ilog]{ilog()}},
\code{\link[=ilog10]{ilog10()}},
\code{\link[=ilog2]{ilog2()}},
\code{\link[=ilogit]{ilogit()}},
\code{\link[=inv_logit]{inv_logit()}},
\code{\link[=invlogit]{invlogit()}},
\code{\link[=log10<-]{log10<-()}},
\code{\link[=log<-]{log<-()}},
\code{\link[=logit]{logit()}},
\code{\link[=logit<-]{logit<-()}},
\code{\link[=phi]{phi()}},
\code{\link[=pow]{pow()}},
\code{\link[=step]{step()}}
}
\concept{translations}
