% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_beta}
\alias{log_lik_beta}
\title{Beta Log-Likelihood}
\usage{
log_lik_beta(x, alpha = 1, beta = 1, tlower = 0, tupper = 1)
}
\arguments{
\item{x}{A vector of values between 0s and 1s.}

\item{alpha}{The first shape parameter of the beta distribution.}

\item{beta}{The second shape parameter of the beta distribution.}

\item{tlower}{A numeric vector of the lower truncation point.}

\item{tupper}{A numeric vector of the upper truncation point.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Beta Log-Likelihood
}
\examples{
log_lik_beta(c(0, 0.5, 0.7, 1), 0.7)
}
\seealso{
Other log_lik_dist:
\code{\link[=log_lik_bern]{log_lik_bern()}},
\code{\link[=log_lik_beta_binom]{log_lik_beta_binom()}},
\code{\link[=log_lik_binom]{log_lik_binom()}},
\code{\link[=log_lik_exp]{log_lik_exp()}},
\code{\link[=log_lik_gamma]{log_lik_gamma()}},
\code{\link[=log_lik_gamma_pois]{log_lik_gamma_pois()}},
\code{\link[=log_lik_gamma_pois_zi]{log_lik_gamma_pois_zi()}},
\code{\link[=log_lik_lnorm]{log_lik_lnorm()}},
\code{\link[=log_lik_multinom]{log_lik_multinom()}},
\code{\link[=log_lik_neg_binom]{log_lik_neg_binom()}},
\code{\link[=log_lik_norm]{log_lik_norm()}},
\code{\link[=log_lik_pois]{log_lik_pois()}},
\code{\link[=log_lik_pois_zi]{log_lik_pois_zi()}},
\code{\link[=log_lik_skewlnorm]{log_lik_skewlnorm()}},
\code{\link[=log_lik_skewnorm]{log_lik_skewnorm()}},
\code{\link[=log_lik_student]{log_lik_student()}},
\code{\link[=log_lik_unif]{log_lik_unif()}}
}
\concept{log_lik_dist}
