% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_lnorm_hurdle}
\alias{log_lik_lnorm_hurdle}
\title{Log-Normal Hurdle Log-Likelihood}
\usage{
log_lik_lnorm_hurdle(x, meanlog = 0, sdlog = 1, prob = 0)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the log scale.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Log-Normal Hurdle Log-Likelihood
}
\examples{
log_lik_lnorm_hurdle(exp(c(-2:2)))
}
\seealso{
Other log_lik_dist: 
\code{\link{log_lik_bern}()},
\code{\link{log_lik_beta_binom}()},
\code{\link{log_lik_binom}()},
\code{\link{log_lik_gamma_pois_zi}()},
\code{\link{log_lik_gamma_pois}()},
\code{\link{log_lik_gamma}()},
\code{\link{log_lik_lnorm}()},
\code{\link{log_lik_neg_binom}()},
\code{\link{log_lik_norm}()},
\code{\link{log_lik_pois_zi}()},
\code{\link{log_lik_pois}()},
\code{\link{log_lik_student}()}
}
\concept{log_lik_dist}
