% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_multinomial}
\alias{log_lik_multinomial}
\title{Multinomial Log-Likelihood}
\usage{
log_lik_multinomial(x, size = 1, prob = 0.5)
}
\arguments{
\item{x}{A matrix of values.}

\item{size}{A non-negative whole numeric vector of the number of trials.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Multinomial Log-Likelihood
}
\examples{
x <- ran_multinomial(10, 5, c(0.2, 0.3, 0.5))
log_lik_multinomial(x, size = 5, prob = c(0.33, 0.33, 0.33))
}
\seealso{
Other log_lik_dist: 
\code{\link{log_lik_bern}()},
\code{\link{log_lik_beta_binom}()},
\code{\link{log_lik_binom}()},
\code{\link{log_lik_gamma}()},
\code{\link{log_lik_gamma_pois}()},
\code{\link{log_lik_gamma_pois_zi}()},
\code{\link{log_lik_lnorm}()},
\code{\link{log_lik_neg_binom}()},
\code{\link{log_lik_norm}()},
\code{\link{log_lik_pois}()},
\code{\link{log_lik_pois_zi}()},
\code{\link{log_lik_skewnorm}()},
\code{\link{log_lik_student}()}
}
\concept{log_lik_dist}
