% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_unif}
\alias{log_lik_unif}
\title{Uniform Log-Likelihood}
\usage{
log_lik_unif(x, min = 0, max = 1)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{min}{A numeric vector of the minimums.}

\item{max}{A numeric vector of the maximums.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Uniform Log-Likelihood
}
\examples{
log_lik_unif(c(0, 0.7, 1))
}
\seealso{
Other log_lik_dist:
\code{\link[=log_lik_bern]{log_lik_bern()}},
\code{\link[=log_lik_beta]{log_lik_beta()}},
\code{\link[=log_lik_beta_binom]{log_lik_beta_binom()}},
\code{\link[=log_lik_binom]{log_lik_binom()}},
\code{\link[=log_lik_exp]{log_lik_exp()}},
\code{\link[=log_lik_gamma]{log_lik_gamma()}},
\code{\link[=log_lik_gamma_pois]{log_lik_gamma_pois()}},
\code{\link[=log_lik_gamma_pois_zi]{log_lik_gamma_pois_zi()}},
\code{\link[=log_lik_lnorm]{log_lik_lnorm()}},
\code{\link[=log_lik_multinom]{log_lik_multinom()}},
\code{\link[=log_lik_neg_binom]{log_lik_neg_binom()}},
\code{\link[=log_lik_norm]{log_lik_norm()}},
\code{\link[=log_lik_pois]{log_lik_pois()}},
\code{\link[=log_lik_pois_zi]{log_lik_pois_zi()}},
\code{\link[=log_lik_skewlnorm]{log_lik_skewlnorm()}},
\code{\link[=log_lik_skewnorm]{log_lik_skewnorm()}},
\code{\link[=log_lik_student]{log_lik_student()}}
}
\concept{log_lik_dist}
