% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{log_odds}
\alias{log_odds}
\title{Log Odds}
\usage{
log_odds(x)
}
\arguments{
\item{x}{A numeric object (vector, matrix or array) of probabilities.}
}
\value{
A numeric object of the the log odds for each probability.
}
\description{
Calculates the log odds for probabilities.
}
\examples{
log_odds(c(0, 0.5, 0.9, 1))
}
\seealso{
Other odds: 
\code{\link{inv_odds}()},
\code{\link{log_odds<-}()},
\code{\link{log_odds_ratio}()},
\code{\link{odds}()},
\code{\link{odds<-}()},
\code{\link{odds_ratio}()}
}
\concept{odds}
