% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{log_odds_ratio}
\alias{log_odds_ratio}
\title{Log-Odds Ratio}
\usage{
log_odds_ratio(x, x2)
}
\arguments{
\item{x}{A numeric object (vector, matrix or array) of probabilities.}

\item{x2}{A second numeric object of probabilities.}
}
\value{
A numeric object of the log odds ratios.
}
\description{
Calculates the log odds ratio for two probabilities.
}
\examples{
log_odds_ratio(0.5, 0.75)
}
\seealso{
Other odds: 
\code{\link{inv_odds}()},
\code{\link{odds<-}()},
\code{\link{odds_ratio}()},
\code{\link{odds}()}
}
\concept{odds}
