% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{log_odds_ratio2}
\alias{log_odds_ratio2}
\title{Log Odds Ratio2}
\usage{
log_odds_ratio2(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
A number.
}
\description{
Calculates the log odds ratio for a vector of two probabilities.
}
\examples{
log_odds_ratio2(c(0.5,0.9))
log_odds_ratio2(c(0.9,0.5))
}
\seealso{
Other odds, fun2: 
\code{\link{odds_ratio2}()}
}
\concept{odds, fun2}
