% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit<-}
\alias{logit<-}
\title{Logistic Transformation}
\usage{
logit(x) <- value
}
\arguments{
\item{x}{An existing object.}

\item{value}{A numeric atomic object of the value to
inverse logistically transform.}
}
\value{
Called for the side effect of updating \code{x}.
}
\description{
Logistic Transformation
}
\details{
A wrapper on \code{stats::plogis(value)}.
}
\examples{
x <- 1
logit(x) <- 0.5
x
}
\seealso{
Other translations:
\code{\link[=exp10]{exp10()}},
\code{\link[=exp2]{exp2()}},
\code{\link[=fabs]{fabs()}},
\code{\link[=ilog]{ilog()}},
\code{\link[=ilog10]{ilog10()}},
\code{\link[=ilog2]{ilog2()}},
\code{\link[=ilogit]{ilogit()}},
\code{\link[=inv_logit]{inv_logit()}},
\code{\link[=invlogit]{invlogit()}},
\code{\link[=log10<-]{log10<-()}},
\code{\link[=log2<-]{log2<-()}},
\code{\link[=log<-]{log<-()}},
\code{\link[=logit]{logit()}},
\code{\link[=phi]{phi()}},
\code{\link[=pow]{pow()}},
\code{\link[=step]{step()}}
}
\concept{translations}
