% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logistic Transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{A numeric atomic object.}
}
\value{
The logistically transformed numeric atomic object.
}
\description{
Logistic transforms a numeric atomic object.
}
\details{
A wrapper on \code{\link[stats:qlogis]{stats::qlogis()}}.
}
\examples{
logit(c(0.25, 0.5, 0.75))
}
\seealso{
Other translations:
\code{\link[=exp10]{exp10()}},
\code{\link[=exp2]{exp2()}},
\code{\link[=fabs]{fabs()}},
\code{\link[=ilog]{ilog()}},
\code{\link[=ilog10]{ilog10()}},
\code{\link[=ilog2]{ilog2()}},
\code{\link[=ilogit]{ilogit()}},
\code{\link[=inv_logit]{inv_logit()}},
\code{\link[=invlogit]{invlogit()}},
\code{\link[=log10<-]{log10<-()}},
\code{\link[=log2<-]{log2<-()}},
\code{\link[=log<-]{log<-()}},
\code{\link[=logit<-]{logit<-()}},
\code{\link[=phi]{phi()}},
\code{\link[=pow]{pow()}},
\code{\link[=step]{step()}}
}
\concept{translations}
