% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{lower}
\alias{lower}
\title{Lower Credible Limit}
\usage{
lower(x, conf_level = 0.95)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}

\item{conf_level}{A numeric scalar between 0 and 1
specifying the confidence level.}
}
\value{
A number.
}
\description{
Calculates the quantile-based lower credible limit.
}
\details{
By default it returns the 95\% credible limit which
corresponds to the 2.5\% quantile.
}
\examples{
lower(as.numeric(0:100))
}
\seealso{
Other summary: 
\code{\link{pvalue}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
