% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims.data.frame}
\alias{ndims.data.frame}
\title{Dimensions of a Data Frame}
\usage{
\method{ndims}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of dimensions.
}
\description{
Gets the number of dimensions of an object.
}
\details{
Always 2L.
}
\examples{
ndims(data.frame())
}
\seealso{
Other {dimensions}: 
\code{\link[universals]{dims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims}()},
\code{\link[universals]{tindex}()}
}
