% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims.default}
\alias{ndims.default}
\title{Number of Dimensions of an Object}
\usage{
\method{ndims}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of dimensions.
}
\description{
Gets the number of dimensions of an object as defined by \code{\link[universals:dims]{dims()}}.
}
\examples{
ndims(character(0))
ndims(1:3)
}
\seealso{
\code{\link[universals:dims]{dims()}}
}
