% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims.matrix}
\alias{ndims.matrix}
\title{Dimensions of a Matrix}
\usage{
\method{ndims}{matrix}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of dimensions.
}
\description{
Gets the number of dimensions of an object.
}
\details{
Always 2L.
}
\examples{
ndims(matrix(1))
}
\seealso{
Other {dimensions}: 
\code{\link[universals]{dims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims}()},
\code{\link[universals]{tindex}()}
}
