% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims-terms.R
\name{npdims_terms.default}
\alias{npdims_terms.default}
\title{Term Number of Parameter Dimensions}
\usage{
\method{npdims_terms}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of integer vectors of the dimensions of each parameter for each term.
}
\description{
Gets the number of dimensions of each parameter for each term.
}
\seealso{
Other {dimensions}: 
\code{\link[universals]{dims}()},
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims_terms}()},
\code{\link[universals]{pdims}()},
\code{\link[universals]{tindex}()}
}
