% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsams.R
\name{nsams.default}
\alias{nsams.default}
\title{Number of MCMC Samples}
\usage{
\method{nsams}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of sample values.
}
\description{
Gets the number of MCMC samples (simulations * terms)
of a Monte Carlo Markov Chain (MCMC) object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}
}
