% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsims.R
\name{nsims.default}
\alias{nsims.default}
\title{Number of Simulations}
\usage{
\method{nsims}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of simulations.
}
\description{
Gets the number of simulations (iterations * chains)
of an MCMC object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars_scalar}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nterms}()}
}
