% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.data.frame}
\alias{numericise.data.frame}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A double numeric atomic matrix object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
If you want to implement a method for a class \code{"foo"}, implement
\code{numericise.foo()}.
}
\examples{
numericise(data.frame(
  logical = c(TRUE, FALSE, NA),
  integer = 1:3,
  numeric = c(4, 10, NA),
  factor = as.factor(c("c", "A", "green"))
))
}
