% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.Date}
\alias{numericise.Date}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{Date}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A double numeric atomic vector object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
If you want to implement a method for a class \code{"foo"}, implement
\code{numericise.foo()}.
}
\examples{
numericise(as.Date("1972-01-01"))
}
