% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.default}
\alias{numericise.default}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A double numeric atomic object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
}
\examples{
as.POSIXlt("1970-01-01", tz = "UTC")
}
