% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.factor}
\alias{numericise.factor}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{factor}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A positive integer numeric atomic vector object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
}
\examples{
numericise(factor(c("c", "a")))
}
