% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.integer}
\alias{numericise.integer}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{integer}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer numeric atomic object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
}
\examples{
numericise(2L)
}
