% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.logical}
\alias{numericise.logical}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{logical}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer numeric atomic object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
If you want to implement a method for a class \code{"foo"}, implement
\code{numericise.foo()}.
}
\examples{
numericise(TRUE)
numericise(matrix(c(TRUE, FALSE), nrow = 2))
}
