% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.matrix}
\alias{numericise.matrix}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{matrix}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A double numeric atomic matrix object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
}
\examples{
numericise(matrix(TRUE))
}
