% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.POSIXct}
\alias{numericise.POSIXct}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{POSIXct}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A double numeric atomic vector object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
If you want to implement a method for a class \code{"foo"}, implement
\code{numericise.foo()}.
}
\examples{
numericise(as.POSIXct("1972-01-01", tz = "UTC"))
}
