% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise}
\alias{numericise}
\alias{numericize}
\title{Numericise (or Numericize)}
\usage{
numericise(x, ...)

numericize(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A numeric atomic object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
\code{numericize()} is an alias for numericise.
If you want to implement a method for a class \code{"foo"}, implement
\code{numericise.foo()}.
}
\examples{
numericise.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to numericise an object of class 'foobar'
}
numericise.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to numericise an object of class 'foobar'
}
}
