% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds<-}
\alias{odds<-}
\title{Inverse Odds Transformation}
\usage{
odds(x) <- value
}
\arguments{
\item{x}{An existing R object.}

\item{value}{A numeric atomic object.}
}
\value{
Called for the side effect of updating \code{x}.
}
\description{
Replaces an object with the inverse odds of value.
}
\examples{
x <- NULL
odds(x) <- 0.5
x
}
\seealso{
Other odds: 
\code{\link{inv_odds}()},
\code{\link{log_odds_ratio}()},
\code{\link{odds_ratio}()},
\code{\link{odds}()}
}
\concept{odds}
