% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Odds Ratio}
\usage{
odds_ratio(x, x2)
}
\arguments{
\item{x}{A numeric object (vector, matrix or array) of probabilities.}

\item{x2}{A second numeric object of probabilities.}
}
\value{
A numeric object of the odds ratios.
}
\description{
Calculates the odds ratio for two probabilities.
}
\examples{
odds_ratio(0.5, 0.75)
}
\seealso{
\code{\link[=odds]{odds()}}
}
