% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds_ratio2}
\alias{odds_ratio2}
\title{Odds Ratio2}
\usage{
odds_ratio2(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
A number.
}
\description{
Calculates the odds ratio for a vector of two probabilities.
}
\examples{
odds_ratio2(c(0.5, 0.9))
odds_ratio2(c(0.9, 0.5))
}
\seealso{
Other odds fun2: 
\code{\link{log_odds_ratio2}()}
}
\concept{odds fun2}
