% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par-pattern.R
\name{par_pattern}
\alias{par_pattern}
\title{Parameter Pattern}
\usage{
par_pattern(ht = FALSE, ds = FALSE)
}
\arguments{
\item{ht}{A flag specifying whether to begin the pattern with "^".}

\item{ds}{A flag specifying whether to end the pattern with "$".}
}
\value{
A string of the regular expression for a parameter name.
}
\description{
Parameter Pattern
}
\examples{
par_pattern()
par_pattern(ht = TRUE)
par_pattern(ds = TRUE)
}
