% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent_change}
\alias{percent_change}
\title{Percent Change}
\usage{
percent_change(x, x2)
}
\arguments{
\item{x}{A numeric object (vector, matrix or array) of non-negative numbers.}

\item{x2}{A second numeric object of non-negative numbers.}
}
\value{
A numeric object of the percent change.
}
\description{
Calculates the percent change for two sets of numbers.
}
\examples{
percent_change(1, 2)
percent_change(2, 1)
}
\seealso{
Other percent: 
\code{\link{percent_difference}()}
}
\concept{percent}
