% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent_change2}
\alias{percent_change2}
\title{Percent Change2}
\usage{
percent_change2(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
A number.
}
\description{
Calculates the percent change for a vector of two non-negative numbers.
}
\examples{
percent_change2(c(1, 2))
percent_change2(c(2, 1))
}
\seealso{
Other percent, fun2: 
\code{\link{percent_difference2}()}
}
\concept{percent, fun2}
