% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent_difference2}
\alias{percent_difference2}
\title{Percent Difference2}
\usage{
percent_difference2(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
A number.
}
\description{
Calculates the percent difference for a vector of two non-negative numbers.
}
\examples{
percent_difference2(c(1, 2))
percent_difference2(c(2, 1))
}
\seealso{
Other percent, fun2: 
\code{\link{percent_change2}()}
}
\concept{percent, fun2}
