% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pextreme.R
\name{pextreme}
\alias{pextreme}
\title{Extreme Probability}
\usage{
pextreme(x)
}
\arguments{
\item{x}{A numeric vector of values between 0 and 1.}
}
\value{
A numeric vector of values between 0 and 1.
}
\description{
Calculates the probability that a cumulative distribution function
probability is at least that extreme.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
pextreme(seq(0, 1, by = 0.1))
}
\seealso{
Other residuals: 
\code{\link{sextreme}()}
}
\concept{residuals}
