% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pkgtemplate_deprecated}
\alias{pkgtemplate_deprecated}
\alias{check_function_template}
\title{Deprecated Functions}
\usage{
check_function_template(x, x_name = NULL)
}
\description{
Deprecated \code{pkgtemplate} functions.
}
\section{Functions}{
\itemize{
\item \code{check_function_template}: Check Function Template

\lifecycle{soft-deprecated}
}}

\examples{

check_function_template("1")
}
\keyword{internal}
