% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Power}
\usage{
pow(x, n)
}
\arguments{
\item{x}{A numeric atomic object of the base.}

\item{n}{A numeric atomic object of the exponent.}
}
\value{
A numeric atomic object of x raised to n.
}
\description{
R equivalent to the power function.
}
\details{
Wrapper on \code{x^n}.
}
\examples{
pow(10, 2)
}
\seealso{
Other translations:
\code{\link[=exp10]{exp10()}},
\code{\link[=exp2]{exp2()}},
\code{\link[=fabs]{fabs()}},
\code{\link[=ilog]{ilog()}},
\code{\link[=ilog10]{ilog10()}},
\code{\link[=ilog2]{ilog2()}},
\code{\link[=ilogit]{ilogit()}},
\code{\link[=inv_logit]{inv_logit()}},
\code{\link[=invlogit]{invlogit()}},
\code{\link[=log10<-]{log10<-()}},
\code{\link[=log2<-]{log2<-()}},
\code{\link[=log<-]{log<-()}},
\code{\link[=logit]{logit()}},
\code{\link[=logit<-]{logit<-()}},
\code{\link[=phi]{phi()}},
\code{\link[=step]{step()}}
}
\concept{translations}
