% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_beta}
\alias{prob_beta}
\title{Beta Cumulative Distribution Function}
\usage{
prob_beta(x, alpha = 1, beta = 1)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{alpha}{The first shape parameter of the beta distribution.}

\item{beta}{The second shape parameter of the beta distribution.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Beta Cumulative Distribution Function
}
\examples{
prob_beta(c(0, 0.5, 1), 2, 3)
}
\seealso{
Other prob_dist:
\code{\link[=prob_bern]{prob_bern()}},
\code{\link[=prob_beta_binom]{prob_beta_binom()}},
\code{\link[=prob_binom]{prob_binom()}},
\code{\link[=prob_exp]{prob_exp()}},
\code{\link[=prob_gamma]{prob_gamma()}},
\code{\link[=prob_gamma_pois]{prob_gamma_pois()}},
\code{\link[=prob_gamma_pois_zi]{prob_gamma_pois_zi()}},
\code{\link[=prob_lnorm]{prob_lnorm()}},
\code{\link[=prob_neg_binom]{prob_neg_binom()}},
\code{\link[=prob_norm]{prob_norm()}},
\code{\link[=prob_pois]{prob_pois()}},
\code{\link[=prob_pois_zi]{prob_pois_zi()}},
\code{\link[=prob_skewlnorm]{prob_skewlnorm()}},
\code{\link[=prob_skewnorm]{prob_skewnorm()}},
\code{\link[=prob_student]{prob_student()}},
\code{\link[=prob_unif]{prob_unif()}}
}
\concept{prob_dist}
