% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_binom}
\alias{prob_binom}
\title{Binomial Cumulative Distribution Function}
\usage{
prob_binom(x, size = 1, prob = 0.5)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{size}{A non-negative whole numeric vector of the number of trials.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of
success.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Binomial Cumulative Distribution Function
}
\examples{
prob_binom(c(0, 1, 2), 2, 0.3)
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
