% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_exp}
\alias{prob_exp}
\title{Exponential Cumulative Distribution Function}
\usage{
prob_exp(x, rate = 1)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{rate}{A non-negative numeric vector of rate.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Exponential Cumulative Distribution Function
}
\examples{
prob_exp(c(0, 1, 2), 2)
}
\seealso{
Other prob_dist:
\code{\link[=prob_bern]{prob_bern()}},
\code{\link[=prob_beta]{prob_beta()}},
\code{\link[=prob_beta_binom]{prob_beta_binom()}},
\code{\link[=prob_binom]{prob_binom()}},
\code{\link[=prob_gamma]{prob_gamma()}},
\code{\link[=prob_gamma_pois]{prob_gamma_pois()}},
\code{\link[=prob_gamma_pois_zi]{prob_gamma_pois_zi()}},
\code{\link[=prob_lnorm]{prob_lnorm()}},
\code{\link[=prob_neg_binom]{prob_neg_binom()}},
\code{\link[=prob_norm]{prob_norm()}},
\code{\link[=prob_pois]{prob_pois()}},
\code{\link[=prob_pois_zi]{prob_pois_zi()}},
\code{\link[=prob_skewlnorm]{prob_skewlnorm()}},
\code{\link[=prob_skewnorm]{prob_skewnorm()}},
\code{\link[=prob_student]{prob_student()}},
\code{\link[=prob_unif]{prob_unif()}}
}
\concept{prob_dist}
