% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_gamma_pois}
\alias{prob_gamma_pois}
\title{Gamma-Poisson Cumulative Distribution Function}
\usage{
prob_gamma_pois(x, lambda = 1, theta = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Gamma-Poisson Cumulative Distribution Function
}
\examples{
prob_gamma_pois(c(0, 1, 2), 1, 1)
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
