% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_gamma_pois_zi}
\alias{prob_gamma_pois_zi}
\title{Zero-Inflated Gamma-Poisson Cumulative Distribution Function}
\usage{
prob_gamma_pois_zi(x, lambda = 1, theta = 0, prob = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of
success.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Zero-Inflated Gamma-Poisson Cumulative Distribution Function
}
\examples{
prob_gamma_pois_zi(c(1, 3, 4), 3, 1, prob = 0.5)
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
