% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_lnorm}
\alias{prob_lnorm}
\title{Log-Normal Cumulative Distribution Function}
\usage{
prob_lnorm(x, meanlog = 0, sdlog = 1)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the
log scale.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Log-Normal Cumulative Distribution Function
}
\examples{
prob_lnorm(10, 0, 2)
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
