% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_neg_binom}
\alias{prob_neg_binom}
\title{Negative Binomial Cumulative Distribution Function}
\usage{
prob_neg_binom(x, lambda = 1, theta = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Negative Binomial Cumulative Distribution Function
}
\examples{
prob_neg_binom(c(0, 1, 2), 2, 1)
}
\seealso{
Other prob_dist:
\code{\link[=prob_bern]{prob_bern()}},
\code{\link[=prob_beta]{prob_beta()}},
\code{\link[=prob_beta_binom]{prob_beta_binom()}},
\code{\link[=prob_binom]{prob_binom()}},
\code{\link[=prob_exp]{prob_exp()}},
\code{\link[=prob_gamma]{prob_gamma()}},
\code{\link[=prob_gamma_pois]{prob_gamma_pois()}},
\code{\link[=prob_gamma_pois_zi]{prob_gamma_pois_zi()}},
\code{\link[=prob_lnorm]{prob_lnorm()}},
\code{\link[=prob_norm]{prob_norm()}},
\code{\link[=prob_pois]{prob_pois()}},
\code{\link[=prob_pois_zi]{prob_pois_zi()}},
\code{\link[=prob_skewnorm]{prob_skewnorm()}},
\code{\link[=prob_student]{prob_student()}},
\code{\link[=prob_unif]{prob_unif()}}
}
\concept{prob_dist}
