% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_norm}
\alias{prob_norm}
\title{Normal Cumulative Distribution Function}
\usage{
prob_norm(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Normal Cumulative Distribution Function
}
\examples{
prob_norm(c(-2:2))
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
