% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_pois}
\alias{prob_pois}
\title{Poisson Cumulative Distribution Function}
\usage{
prob_pois(x, lambda = 1)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{lambda}{A non-negative numeric vector of means.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Poisson Cumulative Distribution Function
}
\examples{
prob_pois(c(1, 3, 4), 3)
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
