% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_skewnorm}
\alias{prob_skewnorm}
\title{Skew Normal Cumulative Distribution Function}
\usage{
prob_skewnorm(x, location = 0, scale = 1, shape = 0, ..., mean, sd)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{location}{A numeric vector of the location parameter.}

\item{scale}{A non-negative numeric vector of the scale parameter.}

\item{shape}{A numeric vector of shape. Negative values result in leftward skew, while positive values result in rightward skew.}

\item{...}{Unused.}

\item{mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of the location parameter.
Described as "a numeric vector of the means" prior to to v. 0.10.1.
Will be removed in a future version.}

\item{sd}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of the scale parameter.
Described as "a non-negative numeric vector of the standard deviations" prior
to v. 0.10.1.
Will be removed in a future version.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Skew Normal Cumulative Distribution Function
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
prob_skewnorm(c(-2:2))
prob_skewnorm(c(-2:2), shape = -2)
prob_skewnorm(c(-2:2), shape = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other prob_dist:
\code{\link[=prob_bern]{prob_bern()}},
\code{\link[=prob_beta]{prob_beta()}},
\code{\link[=prob_beta_binom]{prob_beta_binom()}},
\code{\link[=prob_binom]{prob_binom()}},
\code{\link[=prob_exp]{prob_exp()}},
\code{\link[=prob_gamma]{prob_gamma()}},
\code{\link[=prob_gamma_pois]{prob_gamma_pois()}},
\code{\link[=prob_gamma_pois_zi]{prob_gamma_pois_zi()}},
\code{\link[=prob_lnorm]{prob_lnorm()}},
\code{\link[=prob_neg_binom]{prob_neg_binom()}},
\code{\link[=prob_norm]{prob_norm()}},
\code{\link[=prob_pois]{prob_pois()}},
\code{\link[=prob_pois_zi]{prob_pois_zi()}},
\code{\link[=prob_skewlnorm]{prob_skewlnorm()}},
\code{\link[=prob_student]{prob_student()}},
\code{\link[=prob_unif]{prob_unif()}}
}
\concept{prob_dist}
