% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_student}
\alias{prob_student}
\title{Student's t Cumulative Distribution Function}
\usage{
prob_student(x, mean = 0, sd = 1, theta = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Student's t Cumulative Distribution Function
}
\examples{
prob_student(c(1, 3.5, 4), mean = 1, sd = 2, theta = 1 / 3)
}
\seealso{
Other prob_dist:
\code{\link[=prob_bern]{prob_bern()}},
\code{\link[=prob_beta]{prob_beta()}},
\code{\link[=prob_beta_binom]{prob_beta_binom()}},
\code{\link[=prob_binom]{prob_binom()}},
\code{\link[=prob_exp]{prob_exp()}},
\code{\link[=prob_gamma]{prob_gamma()}},
\code{\link[=prob_gamma_pois]{prob_gamma_pois()}},
\code{\link[=prob_gamma_pois_zi]{prob_gamma_pois_zi()}},
\code{\link[=prob_lnorm]{prob_lnorm()}},
\code{\link[=prob_neg_binom]{prob_neg_binom()}},
\code{\link[=prob_norm]{prob_norm()}},
\code{\link[=prob_pois]{prob_pois()}},
\code{\link[=prob_pois_zi]{prob_pois_zi()}},
\code{\link[=prob_skewlnorm]{prob_skewlnorm()}},
\code{\link[=prob_skewnorm]{prob_skewnorm()}},
\code{\link[=prob_unif]{prob_unif()}}
}
\concept{prob_dist}
