% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_unif}
\alias{prob_unif}
\title{Uniform Cumulative Distribution Function}
\usage{
prob_unif(x, min = 0, max = 1)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{min}{A numeric vector of the minimums.}

\item{max}{A numeric vector of the maximums.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Uniform Cumulative Distribution Function
}
\examples{
prob_unif(c(0, 0.5, 1))
}
\seealso{
Other prob_dist:
\code{\link[=prob_bern]{prob_bern()}},
\code{\link[=prob_beta]{prob_beta()}},
\code{\link[=prob_beta_binom]{prob_beta_binom()}},
\code{\link[=prob_binom]{prob_binom()}},
\code{\link[=prob_exp]{prob_exp()}},
\code{\link[=prob_gamma]{prob_gamma()}},
\code{\link[=prob_gamma_pois]{prob_gamma_pois()}},
\code{\link[=prob_gamma_pois_zi]{prob_gamma_pois_zi()}},
\code{\link[=prob_lnorm]{prob_lnorm()}},
\code{\link[=prob_neg_binom]{prob_neg_binom()}},
\code{\link[=prob_norm]{prob_norm()}},
\code{\link[=prob_pois]{prob_pois()}},
\code{\link[=prob_pois_zi]{prob_pois_zi()}},
\code{\link[=prob_skewnorm]{prob_skewnorm()}},
\code{\link[=prob_student]{prob_student()}}
}
\concept{prob_dist}
