% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-direction.R
\name{probability_direction}
\alias{probability_direction}
\title{Probability of Direction}
\usage{
probability_direction(x, side = "median", threshold = 0, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}

\item{side}{A character vector of length 1 indicating whether to calculate
the directional probability for the left tail (\code{"left"}; \code{x < threshold}),
or the right tail (\code{"right"}; \code{x > threshold}).
Defaults to \code{"median"}, which uses the side of the median of \code{x} via
\code{\link[=direction]{direction()}}.}

\item{threshold}{A number of the threshold value, which is excluded from the
interval for the probability.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number between 0 and 1.
If \code{x} has \code{NA} values but \code{na.rm} is \code{FALSE}, returns \code{NA_real}.
}
\description{
The probability of direction (PD) is the proportion of the (posterior)
distribution above (right) or below (left) a threshold.
}
\details{
By default, the direction is based on the side of the median value, but it
can be specified to measure support for specific hypotheses.
A right-side PD of 0.9 indicates that the interval spanning from the
threshold to infinity has a coverage of 90\%.
Can be used as a measure of certainty in the direction of the estimate
(e.g., positive or negative when using a threshold of 0).
\strong{NOTE:} probability estimates of 0 or 1 are corrected towards 0.5 by adding
or subtracting \code{1 / (length(x) + 1)}, where \code{x} is a vector of MCMC samples.
Ideally, \code{x} should be large enough as to make the correction negligible.
}
\examples{
x <- rnorm(1e6, qnorm(0.05, lower.tail = TRUE))
probability_direction(x, side = "left")
probability_direction(x, side = "right") # = 1 - probability_direction(x, side = "left")
probability_direction(c(0, 0, 1), side = "right") # returns P(X >0) = 1/3 instead of P(X >= 0) = 1
probability_direction(c(1, 1), side = "right") # p = 1 - 1/(n+1)
}
\references{
Makowski, D., Ben-Shachar, M.S., Chen, S.H.A., and Lüdecke, D. 2019. Indices of Effect Existence and Significance in the Bayesian Framework. Front. Psychol. 10: 2767. \doi{10.3389/fpsyg.2019.02767}.
}
\seealso{
Other summary:
\code{\link[=direction]{direction()}},
\code{\link[=directional_information]{directional_information()}},
\code{\link[=kurtosis]{kurtosis()}},
\code{\link[=lower]{lower()}},
\code{\link[=pvalue]{pvalue()}},
\code{\link[=pzeros]{pzeros()}},
\code{\link[=skewness]{skewness()}},
\code{\link[=svalue]{svalue()}},
\code{\link[=upper]{upper()}},
\code{\link[=variance]{variance()}},
\code{\link[=xtr_mean]{xtr_mean()}},
\code{\link[=xtr_median]{xtr_median()}},
\code{\link[=xtr_sd]{xtr_sd()}},
\code{\link[=zeros]{zeros()}},
\code{\link[=zscore]{zscore()}}
}
\concept{summary}
