% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.R
\name{proportional_change2}
\alias{proportional_change2}
\title{Proportional Change2}
\usage{
proportional_change2(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
A number.
}
\description{
Calculates the proportional change for a vector of two non-negative numbers.
}
\examples{
proportional_change2(c(1, 2))
proportional_change2(c(2, 1))
}
\seealso{
Other proportional fun2: 
\code{\link{proportional_difference2}()}
}
\concept{proportional fun2}
