% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.R
\name{proportional_difference}
\alias{proportional_difference}
\title{Proportional Difference}
\usage{
proportional_difference(x, x2)
}
\arguments{
\item{x}{A numeric object (vector, matrix or array) of non-negative numbers.}

\item{x2}{A second numeric object of non-negative numbers.}
}
\value{
A numeric object of the proportional change.
}
\description{
Calculates the proportional difference for two sets of numbers.
}
\examples{
proportional_difference(1, 2)
proportional_difference(2, 1)
}
\seealso{
Other proportional: 
\code{\link{proportional_change}()}
}
\concept{proportional}
