% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{Bayesian P-Value}
\usage{
pvalue(x, threshold = 0, two_sided = TRUE, side = "", na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}

\item{threshold}{A number of the threshold value.}

\item{two_sided}{logical of length 1: calculate two-sided p-values if \code{TRUE} (default) and one-sided if \code{FALSE}. See \code{side}.}

\item{side}{character of length 1: whether to calculate p-values for the left tail (\code{"left"}) or right tail (\code{"right"}).}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number between 0 and 1.
}
\description{
A Bayesian p-value (p) is here defined in terms of the quantile-based
(1-p) * 100\% credible interval (CRI) that
just includes a threshold (Kery and Schaub 2011).
By default a p-value of 0.05 indicates that the 95\% CRI just includes 0.
}
\examples{
x <- rnorm(1e6, qnorm(0.05))
hist(x); abline(v = 0, col = "red", lwd = 2, lty = 2)
pvalue(x) # should be 0.05 * 2
pvalue(x, two_sided = FALSE, side = "left") # should be 0.05
pvalue(x, two_sided = FALSE, side = "right") # should be 0.95
}
\references{
Kery, M., and Schaub, M. 2011.
Bayesian population analysis using WinBUGS: a hierarchical perspective.
Academic Press, Boston. Available from \url{https://www.vogelwarte.ch/en/research/population-biology/book-bpa/}.
}
\seealso{
Other summary: 
\code{\link{kurtosis}()},
\code{\link{lower}()},
\code{\link{pzeros}()},
\code{\link{skewness}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{xtr_median}()},
\code{\link{xtr_sd}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
