% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_binom}
\alias{quant_binom}
\title{Binomial Quantile Function}
\usage{
quant_binom(x, size = 1, prob = 0.5)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{size}{A non-negative whole numeric vector of the number of trials.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of
success.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Binomial Quantile Function
}
\examples{
quant_binom(c(0, 1, 2), 2, 0.3)
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
