% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_exp}
\alias{quant_exp}
\title{Exponential Quantile Function}
\usage{
quant_exp(x, rate = 1)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{rate}{A non-negative numeric vector of rate.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Exponential Quantile Function
}
\examples{
quant_exp(c(0.1, 0.4, 0.6), 2)
}
\seealso{
Other quant_dist:
\code{\link[=quant_bern]{quant_bern()}},
\code{\link[=quant_beta]{quant_beta()}},
\code{\link[=quant_binom]{quant_binom()}},
\code{\link[=quant_gamma]{quant_gamma()}},
\code{\link[=quant_gamma_pois]{quant_gamma_pois()}},
\code{\link[=quant_gamma_pois_zi]{quant_gamma_pois_zi()}},
\code{\link[=quant_lnorm]{quant_lnorm()}},
\code{\link[=quant_neg_binom]{quant_neg_binom()}},
\code{\link[=quant_norm]{quant_norm()}},
\code{\link[=quant_pois]{quant_pois()}},
\code{\link[=quant_pois_zi]{quant_pois_zi()}},
\code{\link[=quant_skewnorm]{quant_skewnorm()}},
\code{\link[=quant_student]{quant_student()}},
\code{\link[=quant_unif]{quant_unif()}}
}
\concept{quant_dist}
