% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_gamma}
\alias{quant_gamma}
\title{Gamma Quantile Function}
\usage{
quant_gamma(x, shape = 1, rate = 1)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{shape}{A non-negative numeric vector of shape.}

\item{rate}{A non-negative numeric vector of rate.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Gamma Quantile Function
}
\examples{
quant_gamma(c(0, 1, 2), 1, 2)
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
