% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_neg_binom}
\alias{quant_neg_binom}
\title{Negative Binomial Quantile Function}
\usage{
quant_neg_binom(x, lambda = 1, theta = 0)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Negative Binomial Quantile Function
}
\examples{
quant_neg_binom(c(0, 1, 2), 2, 1)
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
