% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_norm}
\alias{quant_norm}
\title{Normal Quantile Function}
\usage{
quant_norm(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Normal Quantile Function
}
\examples{
quant_norm(c(-2:2))
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
