% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_pois}
\alias{quant_pois}
\title{Poisson Quantile Function}
\usage{
quant_pois(x, lambda = 1)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{lambda}{A non-negative numeric vector of means.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Poisson Quantile Function
}
\examples{
quant_pois(c(1, 3, 4), 3)
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
