% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_skewnorm}
\alias{quant_skewnorm}
\title{Skew Normal Quantile Function}
\usage{
quant_skewnorm(x, location = 0, scale = 1, shape = 0, ..., mean, sd)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{location}{A numeric vector of the location parameter.}

\item{scale}{A non-negative numeric vector of the scale parameter.}

\item{shape}{A numeric vector of shape. Negative values result in leftward skew, while positive values result in rightward skew.}

\item{...}{Unused.}

\item{mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of the location parameter.
Described as "a numeric vector of the means" prior to to v. 0.10.1.
Will be removed in a future version.}

\item{sd}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of the scale parameter.
Described as "a non-negative numeric vector of the standard deviations" prior
to v. 0.10.1.
Will be removed in a future version.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Skew Normal Quantile Function
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
quant_skewnorm(c(0.1, 0.4, 0.6))
quant_skewnorm(c(0.1, 0.4, 0.6), shape = -2)
quant_skewnorm(c(0.1, 0.4, 0.6), shape = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other quant_dist:
\code{\link[=quant_bern]{quant_bern()}},
\code{\link[=quant_beta]{quant_beta()}},
\code{\link[=quant_binom]{quant_binom()}},
\code{\link[=quant_exp]{quant_exp()}},
\code{\link[=quant_gamma]{quant_gamma()}},
\code{\link[=quant_gamma_pois]{quant_gamma_pois()}},
\code{\link[=quant_gamma_pois_zi]{quant_gamma_pois_zi()}},
\code{\link[=quant_lnorm]{quant_lnorm()}},
\code{\link[=quant_neg_binom]{quant_neg_binom()}},
\code{\link[=quant_norm]{quant_norm()}},
\code{\link[=quant_pois]{quant_pois()}},
\code{\link[=quant_pois_zi]{quant_pois_zi()}},
\code{\link[=quant_skewlnorm]{quant_skewlnorm()}},
\code{\link[=quant_student]{quant_student()}},
\code{\link[=quant_unif]{quant_unif()}}
}
\concept{quant_dist}
