% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_student}
\alias{quant_student}
\title{Student's t Quantile Function}
\usage{
quant_student(x, mean = 0, sd = 1, theta = 0)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Student's t Quantile Function
}
\examples{
quant_student(c(0.1, 0.4, 0.6), mean = 1, sd = 2, theta = 1 / 3)
}
\seealso{
Other quant_dist:
\code{\link[=quant_bern]{quant_bern()}},
\code{\link[=quant_beta]{quant_beta()}},
\code{\link[=quant_binom]{quant_binom()}},
\code{\link[=quant_exp]{quant_exp()}},
\code{\link[=quant_gamma]{quant_gamma()}},
\code{\link[=quant_gamma_pois]{quant_gamma_pois()}},
\code{\link[=quant_gamma_pois_zi]{quant_gamma_pois_zi()}},
\code{\link[=quant_lnorm]{quant_lnorm()}},
\code{\link[=quant_neg_binom]{quant_neg_binom()}},
\code{\link[=quant_norm]{quant_norm()}},
\code{\link[=quant_pois]{quant_pois()}},
\code{\link[=quant_pois_zi]{quant_pois_zi()}},
\code{\link[=quant_skewlnorm]{quant_skewlnorm()}},
\code{\link[=quant_skewnorm]{quant_skewnorm()}},
\code{\link[=quant_unif]{quant_unif()}}
}
\concept{quant_dist}
