% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_unif}
\alias{quant_unif}
\title{Uniform Quantile Function}
\usage{
quant_unif(x, min = 0, max = 1)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{min}{A numeric vector of the minimums.}

\item{max}{A numeric vector of the maximums.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Uniform Quantile Function
}
\examples{
quant_unif(c(0.1, 0.4, 0.6))
}
\seealso{
Other quant_dist:
\code{\link[=quant_bern]{quant_bern()}},
\code{\link[=quant_beta]{quant_beta()}},
\code{\link[=quant_binom]{quant_binom()}},
\code{\link[=quant_exp]{quant_exp()}},
\code{\link[=quant_gamma]{quant_gamma()}},
\code{\link[=quant_gamma_pois]{quant_gamma_pois()}},
\code{\link[=quant_gamma_pois_zi]{quant_gamma_pois_zi()}},
\code{\link[=quant_lnorm]{quant_lnorm()}},
\code{\link[=quant_neg_binom]{quant_neg_binom()}},
\code{\link[=quant_norm]{quant_norm()}},
\code{\link[=quant_pois]{quant_pois()}},
\code{\link[=quant_pois_zi]{quant_pois_zi()}},
\code{\link[=quant_skewlnorm]{quant_skewlnorm()}},
\code{\link[=quant_skewnorm]{quant_skewnorm()}},
\code{\link[=quant_student]{quant_student()}}
}
\concept{quant_dist}
