% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_gamma}
\alias{ran_gamma}
\title{Gamma Random Samples}
\usage{
ran_gamma(n = 1, shape = 1, rate = 1)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to
generate.}

\item{shape}{A non-negative numeric vector of shape.}

\item{rate}{A non-negative numeric vector of rate.}
}
\value{
A numeric vector of the random samples.
}
\description{
Gamma Random Samples
}
\examples{
ran_gamma(10)
}
\seealso{
Other ran_dist:
\code{\link[=ran_bern]{ran_bern()}},
\code{\link[=ran_beta_binom]{ran_beta_binom()}},
\code{\link[=ran_binom]{ran_binom()}},
\code{\link[=ran_gamma_pois]{ran_gamma_pois()}},
\code{\link[=ran_gamma_pois_zi]{ran_gamma_pois_zi()}},
\code{\link[=ran_lnorm]{ran_lnorm()}},
\code{\link[=ran_multinom]{ran_multinom()}},
\code{\link[=ran_neg_binom]{ran_neg_binom()}},
\code{\link[=ran_norm]{ran_norm()}},
\code{\link[=ran_pois]{ran_pois()}},
\code{\link[=ran_pois_zi]{ran_pois_zi()}},
\code{\link[=ran_skewlnorm]{ran_skewlnorm()}},
\code{\link[=ran_skewnorm]{ran_skewnorm()}},
\code{\link[=ran_student]{ran_student()}}
}
\concept{ran_dist}
