% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_multinomial}
\alias{ran_multinomial}
\title{Multinomial Random Samples}
\usage{
ran_multinomial(n = 1, size = 1, prob = 0.5)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to generate.}

\item{size}{An integer}

\item{prob}{A vector of probabilities}
}
\value{
A numeric matrix of the random samples, where each row represents a
draw from the multinomial distribution.
}
\description{
Multinomial Random Samples
}
\examples{
ran_multinomial(10)
}
\seealso{
Other ran_dist: 
\code{\link{ran_bern}()},
\code{\link{ran_beta_binom}()},
\code{\link{ran_binom}()},
\code{\link{ran_gamma}()},
\code{\link{ran_gamma_pois}()},
\code{\link{ran_gamma_pois_zi}()},
\code{\link{ran_lnorm}()},
\code{\link{ran_neg_binom}()},
\code{\link{ran_norm}()},
\code{\link{ran_pois}()},
\code{\link{ran_pois_zi}()},
\code{\link{ran_skewnorm}()},
\code{\link{ran_student}()}
}
\concept{ran_dist}
