% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_neg_binom}
\alias{ran_neg_binom}
\title{Negative Binomial Random Samples}
\usage{
ran_neg_binom(n = 1, lambda = 1, theta = 0)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to
generate.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
A numeric vector of the random samples.
}
\description{
Identical to Gamma-Poisson Random Samples.
}
\examples{
ran_neg_binom(10, theta = 1)
}
\seealso{
Other ran_dist:
\code{\link[=ran_bern]{ran_bern()}},
\code{\link[=ran_beta_binom]{ran_beta_binom()}},
\code{\link[=ran_binom]{ran_binom()}},
\code{\link[=ran_gamma]{ran_gamma()}},
\code{\link[=ran_gamma_pois]{ran_gamma_pois()}},
\code{\link[=ran_gamma_pois_zi]{ran_gamma_pois_zi()}},
\code{\link[=ran_lnorm]{ran_lnorm()}},
\code{\link[=ran_multinom]{ran_multinom()}},
\code{\link[=ran_norm]{ran_norm()}},
\code{\link[=ran_pois]{ran_pois()}},
\code{\link[=ran_pois_zi]{ran_pois_zi()}},
\code{\link[=ran_skewlnorm]{ran_skewlnorm()}},
\code{\link[=ran_skewnorm]{ran_skewnorm()}},
\code{\link[=ran_student]{ran_student()}}
}
\concept{ran_dist}
