% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_skewnorm}
\alias{ran_skewnorm}
\title{Skew Normal Random Samples}
\usage{
ran_skewnorm(n = 1, location = 0, scale = 1, shape = 0, ..., mean, sd)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to
generate.}

\item{location}{A numeric vector of the location parameter.}

\item{scale}{A non-negative numeric vector of the scale parameter.}

\item{shape}{A numeric vector of shape. Negative values result in leftward skew, while positive values result in rightward skew.}

\item{...}{Unused.}

\item{mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of the location parameter.
Described as "a numeric vector of the means" prior to to v. 0.10.1.
Will be removed in a future version.}

\item{sd}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of the scale parameter.
Described as "a non-negative numeric vector of the standard deviations" prior
to v. 0.10.1.
Will be removed in a future version.}
}
\value{
A numeric vector of the random samples.
}
\description{
Skew Normal Random Samples
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
ran_skewnorm(10, shape = -1)
ran_skewnorm(10, shape = 0)
ran_skewnorm(10, shape = 1)
\dontshow{\}) # examplesIf}
}
\seealso{
Other ran_dist:
\code{\link[=ran_bern]{ran_bern()}},
\code{\link[=ran_beta_binom]{ran_beta_binom()}},
\code{\link[=ran_binom]{ran_binom()}},
\code{\link[=ran_gamma]{ran_gamma()}},
\code{\link[=ran_gamma_pois]{ran_gamma_pois()}},
\code{\link[=ran_gamma_pois_zi]{ran_gamma_pois_zi()}},
\code{\link[=ran_lnorm]{ran_lnorm()}},
\code{\link[=ran_multinom]{ran_multinom()}},
\code{\link[=ran_neg_binom]{ran_neg_binom()}},
\code{\link[=ran_norm]{ran_norm()}},
\code{\link[=ran_pois]{ran_pois()}},
\code{\link[=ran_pois_zi]{ran_pois_zi()}},
\code{\link[=ran_skewlnorm]{ran_skewlnorm()}},
\code{\link[=ran_student]{ran_student()}}
}
\concept{ran_dist}
