% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_student}
\alias{ran_student}
\title{Student's t Random Samples}
\usage{
ran_student(n = 1, mean = 0, sd = 1, theta = 0)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to
generate.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
A numeric vector of the random samples.
}
\description{
Student's t Random Samples
}
\examples{
ran_student(10, theta = 1 / 2)
}
\seealso{
Other ran_dist:
\code{\link[=ran_bern]{ran_bern()}},
\code{\link[=ran_beta_binom]{ran_beta_binom()}},
\code{\link[=ran_binom]{ran_binom()}},
\code{\link[=ran_gamma]{ran_gamma()}},
\code{\link[=ran_gamma_pois]{ran_gamma_pois()}},
\code{\link[=ran_gamma_pois_zi]{ran_gamma_pois_zi()}},
\code{\link[=ran_lnorm]{ran_lnorm()}},
\code{\link[=ran_multinom]{ran_multinom()}},
\code{\link[=ran_neg_binom]{ran_neg_binom()}},
\code{\link[=ran_norm]{ran_norm()}},
\code{\link[=ran_pois]{ran_pois()}},
\code{\link[=ran_pois_zi]{ran_pois_zi()}},
\code{\link[=ran_skewlnorm]{ran_skewlnorm()}},
\code{\link[=ran_skewnorm]{ran_skewnorm()}}
}
\concept{ran_dist}
