% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R, R/fill-na.R, R/ndims.R, R/npars.R,
%   R/nsams.R, R/set-pars.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dims}
\alias{fill_na}
\alias{ndims}
\alias{npars}
\alias{nsams}
\alias{set_pars}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{universals}{\code{\link[universals]{dims}}, \code{\link[universals]{fill_na}}, \code{\link[universals]{ndims}}, \code{\link[universals]{npars}}, \code{\link[universals]{nsams}}, \code{\link[universals]{set_pars}}}
}}

