% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{res_gamma_count}
\alias{res_gamma_count}
\title{Gamma-Count Residuals}
\usage{
res_gamma_count(x, lambda = 1, alpha = 1, type = "dev", simulate = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{alpha}{A non-negative numeric vector of the dispersion parameter for the gamma-count distribution.}

\item{type}{A string of the residual type. 'raw' for raw residuals 'dev' for deviance residuals and 'data' for the data.}

\item{simulate}{A flag specifying whether to simulate residuals.}
}
\value{
An numeric vector of the corresponding residuals.
}
\description{
Gamma-Count Residuals
}
\examples{
res_gamma_count(c(0, 1, 2), 1, 1)
}
\seealso{
Other res_dist: 
\code{\link{res_bern}()},
\code{\link{res_beta_binom}()},
\code{\link{res_binom}()},
\code{\link{res_gamma_pois_zi}()},
\code{\link{res_gamma_pois}()},
\code{\link{res_gamma}()},
\code{\link{res_lnorm}()},
\code{\link{res_neg_binom}()},
\code{\link{res_norm}()},
\code{\link{res_pois_zi}()},
\code{\link{res_pois}()},
\code{\link{res_skewnorm}()},
\code{\link{res_student}()}
}
\concept{res_dist}
