% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{res_pois_zi}
\alias{res_pois_zi}
\title{Zero-Inflated Poisson Residuals}
\usage{
res_pois_zi(x, lambda = 1, prob = 0, type = "dev", simulate = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of zero-inflation.}

\item{type}{A string of the residual type. 'raw' for raw residuals 'dev' for
deviance residuals and 'data' for the data.}

\item{simulate}{A flag specifying whether to simulate residuals.}
}
\value{
An numeric vector of the corresponding residuals.
}
\description{
Zero-Inflated Poisson Residuals
}
\examples{
res_pois_zi(c(1, 3, 4), 6, 0.5, type = "raw")
}
\seealso{
Other res_dist:
\code{\link[=res_bern]{res_bern()}},
\code{\link[=res_beta_binom]{res_beta_binom()}},
\code{\link[=res_binom]{res_binom()}},
\code{\link[=res_gamma]{res_gamma()}},
\code{\link[=res_gamma_pois]{res_gamma_pois()}},
\code{\link[=res_gamma_pois_zi]{res_gamma_pois_zi()}},
\code{\link[=res_lnorm]{res_lnorm()}},
\code{\link[=res_multinom]{res_multinom()}},
\code{\link[=res_neg_binom]{res_neg_binom()}},
\code{\link[=res_norm]{res_norm()}},
\code{\link[=res_pois]{res_pois()}},
\code{\link[=res_skewlnorm]{res_skewlnorm()}},
\code{\link[=res_skewnorm]{res_skewnorm()}},
\code{\link[=res_student]{res_student()}}
}
\concept{res_dist}
