% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rope}
\alias{rope}
\title{Region of Practical Equivalence}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{ci_level}{A number > 0 and <= 1 specifying the probability coverage of
the HDI to use.
The default of 1 uses the full posterior.}

\item{quiet}{A flag indicating whether to return warnings.}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A number indicating the estimated proportion of the posterior (or
credible interval) within the region of practical equivalence.
}
\description{
Calculates the proportion of the posterior (or credible interval if
\code{ci_level < 1}) that falls within the region of practical equivalence (ROPE)
to a \code{threshold}.
The ROPE is calculated as \code{threshold + interval}.
Note that the default is not appropriate for all models, since the \code{interval}
is sensitive to unit conversions.
}
\examples{
rope(c(-Inf, -1, -0.1, 0.1, 1, Inf))
rope(rnorm(1e4))
}
\seealso{
\link{xtr_hdi}
}
