% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_beta}
\alias{sens_beta}
\title{Adjust Beta Distribution Parameters for Sensitivity Analyses}
\usage{
sens_beta(alpha, beta, sd_mult = 2)
}
\arguments{
\item{alpha}{The first shape parameter of the Beta distribution.}

\item{beta}{The second shape parameter of the Beta distribution.}

\item{sd_mult}{A non-negative multiplier on the standard deviation of the distribution.}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Beta Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_beta(10, 10, 2)
sens_beta(10, 10, 0.8)
}
\seealso{
Other sens_dist: 
\code{\link{sens_exp}()},
\code{\link{sens_gamma}()},
\code{\link{sens_gamma_pois}()},
\code{\link{sens_gamma_pois_zi}()},
\code{\link{sens_lnorm}()},
\code{\link{sens_neg_binom}()},
\code{\link{sens_norm}()},
\code{\link{sens_pois}()},
\code{\link{sens_pois_zi}()},
\code{\link{sens_skewnorm}()},
\code{\link{sens_student}()}
}
\concept{sens_dist}
