% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_gamma}
\alias{sens_gamma}
\title{Adjust Gamma Distribution Parameters for Sensitivity Analyses}
\usage{
sens_gamma(shape, rate, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\description{
Adjust Gamma Distribution Parameters for Sensitivity Analyses
}
