% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_gamma_pois_zi}
\alias{sens_gamma_pois_zi}
\title{Adjust Zero-Inflated Gamma-Poisson Distribution Parameters for Sensitivity Analyses}
\usage{
sens_gamma_pois_zi(lambda, theta, prob, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Zero-Inflated Gamma-Poisson Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_gamma_pois_zi(10, 0.1, 0.3, 2)
}
