% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_lnorm}
\alias{sens_lnorm}
\title{Adjust Log Normal Distribution Parameters for Sensitivity Analysis}
\usage{
sens_lnorm(meanlog, sdlog, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Log Normal Distribution Parameters for Sensitivity Analysis
}
\examples{
sens_lnorm(0, 1, 2)
sens_lnorm(0, 1, 0.8)
}
