% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_neg_binom}
\alias{sens_neg_binom}
\title{Adjust Negative Binomial Distribution Parameters for Sensitivity Analyses}
\usage{
sens_neg_binom(lambda, theta, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Negative Binomial Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_neg_binom(10, 0.1, 2)
}
