% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_pois}
\alias{sens_pois}
\title{Adjust Poisson Distribution Parameters for Sensitivity Analyses}
\usage{
sens_pois(lambda, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Poisson Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_pois(10, 2)
sens_pois(10, 0.8)
}
