% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_skewnorm}
\alias{sens_skewnorm}
\title{Adjust Skew Normal Distribution Parameters for Sensitivity Analyses}
\usage{
sens_skewnorm(location, scale, shape, scale_mult = 2, ..., mean, sd, sd_mult)
}
\arguments{
\item{location}{A numeric vector of the location parameter.}

\item{scale}{A non-negative numeric vector of the scale parameter.}

\item{shape}{A numeric vector of shape. Negative values result in leftward skew, while positive values result in rightward skew.}

\item{scale_mult}{A non-negative multiplier on the scale of the distribution.}

\item{...}{Unused.}

\item{mean}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of the location parameter.
Described as "a numeric vector of the means" prior to to v. 0.10.1.
Will be removed in a future version.}

\item{sd}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of the scale parameter.
Described as "a non-negative numeric vector of the standard deviations" prior
to v. 0.10.1.
Will be removed in a future version.}

\item{sd_mult}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative multiplier
on the scale of the distribution.}
}
\value{
A named list of the adjusted distribution's parameters.
}
\description{
Expands (\code{scale_mult > 1}) or reduces (\code{scale_mult < 1}) the scale parameter
of the Skew Normal distribution without changing the mean.
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
sens_skewnorm(10, 3, -1, 2)
sens_skewnorm(10, 3, 3, 0.8)
\dontshow{\}) # examplesIf}
}
\seealso{
Other sens_dist:
\code{\link[=sens_beta]{sens_beta()}},
\code{\link[=sens_exp]{sens_exp()}},
\code{\link[=sens_gamma]{sens_gamma()}},
\code{\link[=sens_gamma_pois]{sens_gamma_pois()}},
\code{\link[=sens_gamma_pois_zi]{sens_gamma_pois_zi()}},
\code{\link[=sens_lnorm]{sens_lnorm()}},
\code{\link[=sens_neg_binom]{sens_neg_binom()}},
\code{\link[=sens_norm]{sens_norm()}},
\code{\link[=sens_pois]{sens_pois()}},
\code{\link[=sens_skewlnorm]{sens_skewlnorm()}},
\code{\link[=sens_student]{sens_student()}}
}
\concept{sens_dist}
