% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_student}
\alias{sens_student}
\title{Adjust Student's t Distribution Parameters for Sensitivity Analyses}
\usage{
sens_student(mean, sd, theta, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Student's t Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_student(10, 3, 0.1, 2)
sens_student(10, 3, 0.1, 0.8)
}
