% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sextreme.R
\name{sextreme}
\alias{sextreme}
\title{Extreme Surprisal}
\usage{
sextreme(x, directional = FALSE)
}
\arguments{
\item{x}{A numeric vector of values between 0 and 1.}

\item{directional}{A flag specifying whether probabilities less than 0.5
should be returned as negative values.}
}
\value{
A numeric vector of surprisal values.
}
\description{
Calculates the surprisal (in bits) that a cumulative distribution function
probability is at least that extreme.
}
\details{
Useful for treating the cdf probabilities as residuals
particularly when directional = TRUE.
}
\examples{
sextreme(seq(0.1, 0.9, by = 0.1))
sextreme(seq(0.1, 0.9, by = 0.1), directional = TRUE)
}
\seealso{
Other residuals: 
\code{\link{pextreme}()}
}
\concept{residuals}
