% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewlnorm.R
\name{dskewlnorm}
\alias{dskewlnorm}
\alias{pskewlnorm}
\alias{qskewlnorm}
\alias{rskewlnorm}
\title{Skew-Lognormal Distribution}
\usage{
dskewlnorm(
  x,
  locationlog = 0,
  scalelog = 1,
  shapelog = 0,
  log = FALSE,
  ...,
  meanlog,
  sdlog,
  shape
)

pskewlnorm(
  q,
  locationlog = 0,
  scalelog = 1,
  shapelog = 0,
  ...,
  meanlog,
  sdlog,
  shape
)

qskewlnorm(
  p,
  locationlog = 0,
  scalelog = 1,
  shapelog = 0,
  ...,
  meanlog,
  sdlog,
  shape
)

rskewlnorm(
  n = 1,
  locationlog = 0,
  scalelog = 1,
  shapelog = 0,
  ...,
  meanlog,
  sdlog,
  shape
)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{locationlog}{A numeric vector of location parameters of \code{log(x)}.}

\item{scalelog}{A non-negative numeric vector of scale parameters of \code{log(x)}.}

\item{shapelog}{A numeric vector of shape parameters of \code{log(x)}.
Negative values result in leftward skew, while positive values result in rightward skew.}

\item{log}{A flag specifying whether to return the log-transformed value.}

\item{...}{Unused.}

\item{meanlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of location parameters of \code{log(x)}.
Described as "a numeric vector of the means on the log scale" prior to v.
0.10.1.
Will be removed in a future version.}

\item{sdlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of scale parameters of
\code{log(x)}.
Described as "a non-negative numeric vector of the standard deviations on the
log scale" prior to v. 0.10.1.
Will be removed in a future version.}

\item{shape}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of shape parameters of \code{log(x)}.
Will be removed in a future version.}

\item{q}{A vector of quantiles.}

\item{p}{A numeric vector of probabilities.}

\item{n}{A non-negative whole number of the number of random samples to
generate.}
}
\value{
\code{dskewlnorm} gives the density, \code{pskewlnorm} gives the distribution function, \code{qskewlnorm} gives the quantile function, and \code{rskewlnorm} generates random deviates.
\code{pskewlnorm} and \code{qskewlnorm} use the lower tail probability.
}
\description{
The skew-lognormal distribution of a random variable whose natural logarithm
follows a \link[=dskewnorm]{Skew-Normal} distribution with location \code{locationlog},
scale \code{scalelog} and \code{shapelog}.
It reduces to the Log-Normal distribution when \code{shapelog = 0}.
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
dskewlnorm(x = 1:5, locationlog = 0, scalelog = 1, shapelog = 0.1)
dskewlnorm(x = 1:5, locationlog = 0, scalelog = 1, shapelog = -1)
qskewlnorm(p = c(0.1, 0.4), locationlog = 0, scalelog = 1, shapelog = 0.1)
qskewlnorm(p = c(0.1, 0.4), locationlog = 0, scalelog = 1, shapelog = -1)
pskewlnorm(q = 1:5, locationlog = 0, scalelog = 1, shapelog = 0.1)
pskewlnorm(q = 1:5, locationlog = 0, scalelog = 1, shapelog = -1)
rskewlnorm(n = 3, locationlog = 0, scalelog = 1, shapelog = 0.1)
rskewlnorm(n = 3, locationlog = 0, scalelog = 1, shapelog = -1)
\dontshow{\}) # examplesIf}
}
\concept{skewlnorm}
