% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric object of MCMC values.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number.
}
\description{
Skewness
}
\examples{
skewness(1:10)
}
\seealso{
Other summary:
\code{\link[=kurtosis]{kurtosis()}},
\code{\link[=lower]{lower()}},
\code{\link[=pvalue]{pvalue()}},
\code{\link[=pzeros]{pzeros()}},
\code{\link[=svalue]{svalue()}},
\code{\link[=upper]{upper()}},
\code{\link[=variance]{variance()}},
\code{\link[=xtr_mean]{xtr_mean()}},
\code{\link[=xtr_median]{xtr_median()}},
\code{\link[=xtr_sd]{xtr_sd()}},
\code{\link[=zeros]{zeros()}},
\code{\link[=zscore]{zscore()}}
}
\concept{summary}
