% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.R
\name{upper}
\alias{upper}
\title{Upper Credible Limit}
\usage{
upper(x, conf_level = 0.95, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}

\item{conf_level}{A numeric scalar between 0 and 1
specifying the confidence level.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number.
}
\description{
Calculates the quantile-based upper credible limit.
}
\details{
By default it returns the 95\% credible limit which
corresponds to the 97.5\% quantile.
}
\examples{
upper(as.numeric(0:100))
}
\seealso{
Other summary: 
\code{\link{kurtosis}()},
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{pzeros}()},
\code{\link{skewness}()},
\code{\link{svalue}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{xtr_median}()},
\code{\link{xtr_sd}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
