% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_function_template}
\alias{vld_function_template}
\title{Validate Function Template}
\usage{
vld_function_template(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the condition was met.
}
\description{
A vld function template.
}
\details{
Validates character using

\code{is.character(x)}.
}
\examples{
vld_function_template("1")
vld_function_template(1)
}
\seealso{
\code{\link[=chk_function_template]{chk_function_template()}}
}
