% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{xtr_ci}
\alias{xtr_ci}
\title{Credible Intervals}
\usage{
xtr_ci(x, level = 0.95, ..., type = "HDI", na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{level}{A number > 0 and <= 1 (inclusive) specifying the probability
coverage of the interval.}

\item{...}{Currently unused.}

\item{type}{A string indicating which type of CI to return.
Currently allows Highest Density Intervals (\code{"HDI"}; default) and
Equal-Tailed Intervals (\code{"ETI"}).}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A data frame of the \code{lower} and \code{upper} limits for the credible interval.
Note that the interval is not guaranteed to be one-sided or two-sided.
}
\description{
Calculates Bayesian credible intervals (CI) using one of the available methods:
\itemize{
\item \code{"HDI"}: highest density interval (see \code{\link[=xtr_ci_hdi]{xtr_ci_hdi()}}),
\item \code{"ETI"}: equal tailed intervals (see \code{\link[=xtr_ci_eti]{xtr_ci_eti()}}).
}
}
\examples{
xtr_ci(rnorm(1e4), type = "HDI")
}
\seealso{
\link[mcmcr:coef]{mcmcr::coef}
}
