% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_hdi.R
\name{xtr_ci_hdi}
\alias{xtr_ci_hdi}
\title{Highest Density Interval}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{level}{A number between 0 and 1 (exclusive) specifying the probability
coverage of the HDI.}

\item{quiet}{A flag indicating whether to return warnings.}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of the \code{lower} and \code{upper} limits for the credible interval.
Note that the interval is not guaranteed to be one-sided or two-sided.
}
\description{
Calculates Bayesian credible intervals using the highest density
interval (HDI), i.e., the narrowest CI with the specified minimum coverage.
}
\examples{
xtr_ci_hdi(rnorm(1e4))
}
\seealso{
\code{\link[=xtr_ci]{xtr_ci()}}, \code{\link[=xtr_ci_eti]{xtr_ci_eti()}}, and \code{\link[=xtr_ci_norm]{xtr_ci_norm()}}
}
