% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_norm.R
\name{xtr_ci_norm}
\alias{xtr_ci_norm}
\title{Normal Approximation Intervals}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{level}{A number between 0 and 1 (exclusive) specifying the probability
coverage of the interval.}

\item{quiet}{A flag indicating whether to return warnings.}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A data frame of the \code{lower} and \code{upper} limits for the credible interval.
Note that the interval is not guaranteed to be one-sided or two-sided.
}
\description{
Calculates equal-tailed credible interval using using a normal (i.e.,
Gaussian) approximation of the distribution.
The intervals may be similar to those produced by \code{xtr_ci_eti()} when the
distributions of the data are symmetric.
Returns \code{NA_real_} if any elements of \code{x} are infinite.
}
\examples{
xtr_ci_norm(rnorm(1e4))
}
\seealso{
\link{xtr_ci}, \link{xtr_ci_eti}, and \link{xtr_ci_hdi}
}
