% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{xtr_hdi}
\alias{xtr_hdi}
\title{Highest Density Interval}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{level}{A number between 0 and 1 (exclusive) specifying the probability
coverage of the HDI.}

\item{quiet}{A flag indicating whether to return warnings.}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A data frame of the \code{lower} and \code{upper} limits for the credible interval.
Note that the interval is not guaranteed to be one-sided or two-sided.
}
\description{
Calculates credible intervals using the empirical highest density interval
(HDI), i.e., the narrowest CI with the specified minimum coverage.
}
\examples{
xtr_hdi(rnorm(1e4))
}
\seealso{
\link[mcmcr:coef]{mcmcr::coef}
}
