% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{xtr_mean}
\alias{xtr_mean}
\title{Mean}
\usage{
xtr_mean(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric object of MCMC values.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number.
}
\description{
Mean
}
\examples{
xtr_mean(1:10)
}
\seealso{
Other summary:
\code{\link[=direction]{direction()}},
\code{\link[=directional_information]{directional_information()}},
\code{\link[=kurtosis]{kurtosis()}},
\code{\link[=lower]{lower()}},
\code{\link[=probability_direction]{probability_direction()}},
\code{\link[=pvalue]{pvalue()}},
\code{\link[=pzeros]{pzeros()}},
\code{\link[=skewness]{skewness()}},
\code{\link[=svalue]{svalue()}},
\code{\link[=upper]{upper()}},
\code{\link[=variance]{variance()}},
\code{\link[=xtr_median]{xtr_median()}},
\code{\link[=xtr_sd]{xtr_sd()}},
\code{\link[=zeros]{zeros()}},
\code{\link[=zscore]{zscore()}}
}
\concept{summary}
