% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_bern}
\alias{dev_bern}
\title{Bernoulli Deviance}
\usage{
dev_bern(x, prob, residual = FALSE)
}
\arguments{
\item{x}{A logical vector.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}

\item{residual}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Bernoulli Deviance
}
\examples{
dev_bern(c(TRUE, FALSE), 0.7, residual = TRUE)
}
\seealso{
Other devdist: 
\code{\link{dev_binom}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois}()}
}
\concept{devdist}
