% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_binom}
\alias{dev_binom}
\title{Binomial Deviance}
\usage{
dev_binom(x, size, prob, residual = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{size}{A non-negative whole numeric vector of the number of trials.}

\item{prob}{A numeric vector of values between 0 and 1 o the probability of success of each trial.}

\item{residual}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Binomial Deviance
}
\examples{
dev_binom(c(0, 1, 2), 2, 0.3, residual = TRUE)
}
\seealso{
Other devdist: 
\code{\link{dev_bern}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois}()}
}
\concept{devdist}
