% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_binom}
\alias{dev_binom}
\title{Binomial Deviances}
\usage{
dev_binom(x, size = 1, prob = 0.5, res = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{size}{A non-negative whole numeric vector of the number of trials.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Binomial Deviances
}
\examples{
dev_binom(c(0, 1, 2), 2, 0.3)
}
\seealso{
Other dev_dist: 
\code{\link{dev_bern}()},
\code{\link{dev_beta_binom}()},
\code{\link{dev_gamma_count}()},
\code{\link{dev_gamma_pois}()},
\code{\link{dev_gamma}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_neg_binom}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois_zi}()},
\code{\link{dev_pois}()},
\code{\link{dev_skewnorm}()},
\code{\link{dev_student}()}
}
\concept{dev_dist}
