% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_gamma_pois}
\alias{dev_gamma_pois}
\title{Gamma-Poisson Deviances}
\usage{
dev_gamma_pois(x, lambda, theta, res = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector.}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance ress.
}
\description{
Gamma-Poisson Deviances
}
\examples{
dev_gamma_pois(c(1,3.5,4), 3, 2)
}
\seealso{
Other devdist: 
\code{\link{dev_bern}()},
\code{\link{dev_binom}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois}()}
}
\concept{devdist}
