% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_gamma_pois}
\alias{dev_gamma_pois}
\title{Gamma-Poisson Deviances}
\usage{
dev_gamma_pois(x, lambda = 1, theta = 0, res = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture models (student, gamma-Poisson, beta-binomial, and underdispersed Poisson).}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Gamma-Poisson Deviances
}
\examples{
dev_gamma_pois(c(1,3.5,4), 3, 2)
}
\seealso{
Other dev_dist: 
\code{\link{dev_bern}()},
\code{\link{dev_beta_binom}()},
\code{\link{dev_binom}()},
\code{\link{dev_gamma}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_neg_binom}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois_zi}()},
\code{\link{dev_pois}()},
\code{\link{dev_skewnorm}()},
\code{\link{dev_student}()},
\code{\link{dev_upois}()}
}
\concept{dev_dist}
