% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_lnorm}
\alias{dev_lnorm}
\title{Log-Normal Deviances}
\usage{
dev_lnorm(x, meanlog = 0, sdlog = 1, res = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the log scale.}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Log-Normal Deviances
}
\examples{
dev_lnorm(exp(-2:2))
}
\seealso{
Other dev_dist: 
\code{\link{dev_bern}()},
\code{\link{dev_binom}()},
\code{\link{dev_gamma_pois}()},
\code{\link{dev_neg_binom}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois}()}
}
\concept{dev_dist}
