% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_norm}
\alias{dev_norm}
\title{Normal Deviance}
\usage{
dev_norm(x, mean = 0, sd = 1, residual = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{residual}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Normal Deviance
}
\examples{
dev_norm(c(-2:2), residual = TRUE)
}
\seealso{
Other devdist: 
\code{\link{dev_bern}()},
\code{\link{dev_binom}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_pois}()}
}
\concept{devdist}
