% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_pois_zi}
\alias{dev_pois_zi}
\title{Zero-Inflated Poisson Deviances}
\usage{
dev_pois_zi(x, lambda, prob = 0, res = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Zero-Inflated Poisson Deviances
}
\examples{
dev_pois(c(1,3.5,4), 3)
}
\seealso{
Other dev_dist: 
\code{\link{dev_bern}()},
\code{\link{dev_beta_binom}()},
\code{\link{dev_binom}()},
\code{\link{dev_gamma_count}()},
\code{\link{dev_gamma_pois}()},
\code{\link{dev_gamma}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_neg_binom}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois}()},
\code{\link{dev_skewnorm}()},
\code{\link{dev_student}()}
}
\concept{dev_dist}
