% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev_skewnorm}
\alias{dev_skewnorm}
\title{Skew Normal Deviances}
\usage{
dev_skewnorm(x, mean = 0, sd = 1, shape = 0, res = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{shape}{A numeric vector of shape.}

\item{res}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Skew Normal Deviances
}
\examples{
dev_skewnorm(c(-2:2))
dev_skewnorm(-2:2, 0, 1, 5)
dev_skewnorm(-2:2, 0, 1, 5, res = TRUE)
}
\seealso{
Other dev_dist: 
\code{\link{dev_bern}()},
\code{\link{dev_beta_binom}()},
\code{\link{dev_binom}()},
\code{\link{dev_gamma_pois}()},
\code{\link{dev_gamma}()},
\code{\link{dev_lnorm}()},
\code{\link{dev_neg_binom}()},
\code{\link{dev_norm}()},
\code{\link{dev_pois_zi}()},
\code{\link{dev_pois}()},
\code{\link{dev_student}()},
\code{\link{dev_upois}()}
}
\concept{dev_dist}
