% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-residuals.R
\name{devbern}
\alias{devbern}
\title{Bernoulli Deviance Residual}
\usage{
devbern(x, prob)
}
\arguments{
\item{x}{A logical vector.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of TRUE.}
}
\value{
An numeric vector of the corresponding deviance residuals.
}
\description{
Bernoulli Deviance Residual
}
\examples{
devbern(c(TRUE, FALSE), 0.7)
}
\seealso{
Other devdist: 
\code{\link{devbinom}()},
\code{\link{devlnorm}()},
\code{\link{devnorm}()},
\code{\link{devpois}()}
}
\concept{devdist}
