% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-residuals.R
\name{devlnorm}
\alias{devlnorm}
\title{Log-Normal Deviance Residual}
\usage{
devlnorm(x, meanlog = 0, sdlog = 1)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the log scale.}
}
\value{
An numeric vector of the corresponding deviance residuals.
}
\description{
Log-Normal Deviance Residual
}
\examples{
devlnorm(exp(-2:2))
}
\seealso{
Other devdist: 
\code{\link{devbern}()},
\code{\link{devbinom}()},
\code{\link{devnorm}()},
\code{\link{devpois}()}
}
\concept{devdist}
