% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance-residuals.R
\name{devnorm}
\alias{devnorm}
\title{Normal Deviance Residual}
\usage{
devnorm(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}
}
\value{
An numeric vector of the corresponding deviance residuals.
}
\description{
Normal Deviance Residual
}
\examples{
devnorm(c(-2:2))
}
\seealso{
Other devdist: 
\code{\link{devbern}()},
\code{\link{devbinom}()},
\code{\link{devlnorm}()},
\code{\link{devpois}()}
}
\concept{devdist}
