% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{devpois}
\alias{devpois}
\title{Poisson Deviance}
\usage{
devpois(x, lambda, residual = FALSE)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}

\item{residual}{A flag specifying whether to return the deviance residual as opposed to the deviance.}
}
\value{
An numeric vector of the corresponding deviances or deviance residuals.
}
\description{
Poisson Deviance
}
\examples{
devpois(c(1,3.5,4), 3, residual = TRUE)
}
\seealso{
Other devdist: 
\code{\link{devbern}()},
\code{\link{devbinom}()},
\code{\link{devlnorm}()},
\code{\link{devnorm}()}
}
\concept{devdist}
