% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.data.frame}
\alias{dims.data.frame}
\title{Dimensions of a data frame}
\usage{
\method{dims}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object. See the individual methods for specifics.}

\item{...}{Other arguments passed to methods.}
}
\description{
Dimensions of a data frame
}
\examples{
dims(data.frame())
dims(data.frame(x = 1:3))
}
\seealso{
\code{\link[universals]{dims}()}
}
