% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.data.frame}
\alias{dims.data.frame}
\title{Dimensions of a Data Frame}
\usage{
\method{dims}{data.frame}(x, ...)
}
\arguments{
\item{x}{A data.frame.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
Unlike \code{base::dim()}, dims works as intuitively expected
for vectors and data.frames.
}
\examples{
dims(data.frame())
dims(data.frame(x = 1:3))
}
\seealso{
\code{\link[base:dim]{base::dim()}}
}
