% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.default}
\alias{dims.default}
\title{Dimensions}
\usage{
\method{dims}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
}
\details{
Unlike \code{base::dim()}, dims returns the length of an atomic vector.
}
\examples{
dims(character(0))
dims(1:3)
}
\seealso{
\code{\link[base:dim]{base::dim()}}

Other dimensions: 
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims}()}
}
