% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.factor}
\alias{dims.factor}
\title{Dimensions of a Factor}
\usage{
\method{dims}{factor}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
Unlike \code{base::dim()}, dims works as intuitively expected
for vectors and data.frames.
}
\examples{
dims(factor("a"))
}
\seealso{
\code{\link[base:dim]{base::dim()}}

Other {dimensions}: 
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims}()}
}
