% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{exp2}
\alias{exp2}
\title{Exponential transformation of 2^x}
\usage{
exp2(x)
}
\arguments{
\item{x}{An numeric atomic object.}
}
\value{
A numeric atomic object.
}
\description{
Inverse of log2(x)
}
\details{
Returns the value of 2^x
}
\examples{
x <- 5
exp2(x) # 32
2^x # 32
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{fabs}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
