% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{exp2}
\alias{exp2}
\title{Exponential Transformation of Base 2}
\usage{
exp2(x)
}
\arguments{
\item{x}{An numeric atomic object.}
}
\value{
A numeric atomic object with the value of 2^x.
}
\description{
Returns the transformation of 2^x.
}
\examples{
x <- c(5, 10.5)
exp2(x)
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{fabs}()},
\code{\link{ilog}()},
\code{\link{ilog10}()},
\code{\link{ilog2}()},
\code{\link{ilogit}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log10<-}()},
\code{\link{log2<-}()},
\code{\link{log<-}()},
\code{\link{logit}()},
\code{\link{logit<-}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
