% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras-package.R
\docType{package}
\name{extras-package}
\alias{extras}
\alias{extras-package}
\title{extras: Extra Functions for Data Modeling}
\description{
Extra functions used by data modeling packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/extras}
  \item Report bugs at \url{https://github.com/poissonconsulting/extras/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Evan Amies-Galonski \email{evan@poissonconsulting.ca} (\href{https://orcid.org/-3188}{ORCID}) [contributor]
  \item Ayla Pearson (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
