% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras-package.R
\docType{package}
\name{extras-package}
\alias{extras}
\alias{extras-package}
\title{extras: Extra Functions for Data Modeling}
\description{
Extra functions for data modeling.
    It includes functions to numericise objects; 
    simple methods for S3 generics in the universals package;
    and R translations of common JAGS and TMB functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/extras}
  \item Report bugs at \url{https://github.com/poissonconsulting/extras/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
