% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras-package.R
\docType{package}
\name{extras-package}
\alias{extras}
\alias{extras-package}
\title{extras: Basic Functions for Bayesian Analyses}
\description{
Provides basic functions for Bayesian analyses
    including implementations of simple S3 generics; 
    functions to summarise MCMC samples.
    and R translations of a handful of common functions 
    from modeling languages such as BUGS/JAGS and C++.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/extras}
  \item Report bugs at \url{https://github.com/poissonconsulting/extras/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
