% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras-package.R
\docType{package}
\name{extras-package}
\alias{extras}
\alias{extras-package}
\title{extras: Helper Functions for Bayesian Analyses}
\description{
Functions to 'numericise' 'R' objects (coerce to numeric objects)
    and summarise 'MCMC' (Monte Carlo Markov Chain) samples
    as well as 'R' translations of 'BUGS' (Bayesian Using Gibbs Sampling)
    and 'JAGS' (Just Another Gibbs Sampler) functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/extras/}
  \item \url{https://github.com/poissonconsulting/extras}
  \item Report bugs at \url{https://github.com/poissonconsulting/extras/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
