% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabs.R
\name{fabs}
\alias{fabs}
\title{Absolute}
\usage{
fabs(x)
}
\arguments{
\item{x}{An existing R object.}
}
\description{
Computes the absolute value of \code{x}.
Used in TMB as replacement for \code{abs()} which is seemingly ambiguous.
}
\details{
A wrapper on \code{\link{abs}}\verb{()}.
}
\examples{
fabs(c(0, -1, 2))
}
\seealso{
Other translations: 
\code{\link{ilogit}()},
\code{\link{invlogit}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{odds}()},
\code{\link{phi}()},
\code{\link{pow}()}
}
\concept{translations}
