% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-all.R
\name{fill_all}
\alias{fill_all}
\title{Fill All Values}
\usage{
fill_all(x, value, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace all values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills all of an object's (missing and non-missing) values while preserving the object's dimensionality and class.
}
