% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{ilog}
\alias{ilog}
\title{Inverse Log Transformation}
\usage{
ilog(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A numeric atomic object.
}
\description{
Inverse log transforms a numeric atomic object.
}
\details{
A wrapper on \code{\link{exp}}\code{(value)}.
}
\examples{
x <- 1
ilog(x)
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{exp2}()},
\code{\link{fabs}()},
\code{\link{ilog10}()},
\code{\link{ilog2}()},
\code{\link{ilogit}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log10<-}()},
\code{\link{log2<-}()},
\code{\link{log<-}()},
\code{\link{logit}()},
\code{\link{logit<-}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
