% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog10}
\alias{ilog10}
\title{Inverse Log Base 10 Transformation}
\usage{
ilog10(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A numeric atomic object.
}
\description{
Inverse log transforms a numeric atomic object with base 10.
}
\details{
A wrapper on \code{\link{exp10}}\code{(value)}.
}
\examples{
x <- c(2, 4.5)
ilog10(x)
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{exp2}()},
\code{\link{fabs}()},
\code{\link{ilog}()},
\code{\link{ilog2}()},
\code{\link{ilogit}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log10<-}()},
\code{\link{log2<-}()},
\code{\link{log<-}()},
\code{\link{logit}()},
\code{\link{logit<-}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
