% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilogit.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logistic Transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A numeric atomic object.}
}
\value{
A numeric atomic object.
}
\description{
Inverse logistically transforms a numeric atomic object.
}
\details{
A wrapper on \code{\link[stats:Logistic]{stats::plogis()}}.
}
\examples{
invlogit(c(-1, 0, 5))
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{exp2}()},
\code{\link{fabs}()},
\code{\link{ilog10}()},
\code{\link{ilog2}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{inv_logit}()},
\code{\link{log10<-}()},
\code{\link{log2<-}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
