% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric object of MCMC values.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number.
}
\description{
Kurtosis
}
\examples{
kurtosis(1:10)
}
\seealso{
Other summary: 
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{skewness}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
