% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log<-}
\alias{log<-}
\title{Log Transformation}
\usage{
log(x) <- value
}
\arguments{
\item{x}{An existing object.}

\item{value}{A numeric atomic object of the value to exponentiate.}
}
\description{
Log Transformation
}
\details{
A wrapper on \code{exp(value)}.
}
\examples{
x <- NULL
log(x) <- 0.5
x
}
\seealso{
Other {translations}: 
\code{\link{ilogit}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{phi}()},
\code{\link{pow}()}
}
\concept{{translations}}
