% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log<-}
\alias{log<-}
\title{Log Transformation}
\usage{
log(x) <- value
}
\arguments{
\item{x}{An existing R object.}

\item{value}{A numeric atomic object.}
}
\value{
Called for the side effect of updating \code{x}.
}
\description{
Replaces a object with the exponent of value.
}
\details{
A wrapper on \code{\link{exp}}\code{(value)}.
}
\examples{
x <- NULL
log(x) <- 0.5
x
}
\seealso{
Other translations: 
\code{\link{ilogit}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{phi}()},
\code{\link{pow}()}
}
\concept{translations}
