% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log10<-}
\alias{log10<-}
\title{Log Base 10 Transformation}
\usage{
log10(x) <- value
}
\arguments{
\item{x}{An object.}

\item{value}{A numeric atomic object.}
}
\value{
Called for the side effect of updating \code{x}.
}
\description{
Replaces a object with the base 10 exponent of value.
}
\details{
A wrapper on \code{\link{exp10}}\code{(value)}.
}
\examples{
x <- NULL
log10(x) <- c(0.5, 5)
x
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{exp2}()},
\code{\link{fabs}()},
\code{\link{ilog10}()},
\code{\link{ilog2}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log2<-}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
