% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_binom}
\alias{log_lik_binom}
\title{Binomial Log-Likelihood}
\usage{
log_lik_binom(x, size = 1, prob = 0.5)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{size}{A non-negative whole numeric vector of the number of trials.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Binomial Log-Likelihood
}
\examples{
log_lik_binom(c(0, 1, 2), 2, 0.3)
}
\seealso{
Other log_lik_dist: 
\code{\link{log_lik_bern}()},
\code{\link{log_lik_gamma_pois}()},
\code{\link{log_lik_lnorm}()},
\code{\link{log_lik_neg_binom}()},
\code{\link{log_lik_norm}()},
\code{\link{log_lik_pois}()}
}
\concept{log_lik_dist}
