% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_exp}
\alias{log_lik_exp}
\title{Exponential Log-Likelihood}
\usage{
log_lik_exp(x, mean = 1)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{mean}{A numeric vector of the means.}
}
\value{
A numeric vector of the corresponding log-likelihoods.
}
\description{
The exponential distribution is parameterized in terms of the scale.
}
\examples{
log_lik_exp(c(0, 1, 2), 2)
}
\seealso{
Other log_lik_dist: 
\code{\link{log_lik_bern}()},
\code{\link{log_lik_beta_binom}()},
\code{\link{log_lik_binom}()},
\code{\link{log_lik_gamma}()},
\code{\link{log_lik_gamma_pois}()},
\code{\link{log_lik_gamma_pois_zi}()},
\code{\link{log_lik_lnorm}()},
\code{\link{log_lik_neg_binom}()},
\code{\link{log_lik_norm}()},
\code{\link{log_lik_pois}()},
\code{\link{log_lik_pois_zi}()},
\code{\link{log_lik_skewnorm}()},
\code{\link{log_lik_student}()}
}
\concept{log_lik_dist}
