% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_lnorm}
\alias{log_lik_lnorm}
\title{Log-Normal Log-Likelihood}
\usage{
log_lik_lnorm(x, meanlog = 0, sdlog = 1, tlower = 0, tupper = Inf)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the
log scale.}

\item{tlower}{A numeric vector of the lower truncation point.}

\item{tupper}{A numeric vector of the upper truncation point.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Log-Normal Log-Likelihood
}
\examples{
log_lik_lnorm(10, 0, 2)
}
\seealso{
Other log_lik_dist:
\code{\link[=log_lik_bern]{log_lik_bern()}},
\code{\link[=log_lik_beta]{log_lik_beta()}},
\code{\link[=log_lik_beta_binom]{log_lik_beta_binom()}},
\code{\link[=log_lik_binom]{log_lik_binom()}},
\code{\link[=log_lik_exp]{log_lik_exp()}},
\code{\link[=log_lik_gamma]{log_lik_gamma()}},
\code{\link[=log_lik_gamma_pois]{log_lik_gamma_pois()}},
\code{\link[=log_lik_gamma_pois_zi]{log_lik_gamma_pois_zi()}},
\code{\link[=log_lik_neg_binom]{log_lik_neg_binom()}},
\code{\link[=log_lik_norm]{log_lik_norm()}},
\code{\link[=log_lik_pois]{log_lik_pois()}},
\code{\link[=log_lik_pois_zi]{log_lik_pois_zi()}},
\code{\link[=log_lik_skewlnorm]{log_lik_skewlnorm()}},
\code{\link[=log_lik_skewnorm]{log_lik_skewnorm()}},
\code{\link[=log_lik_student]{log_lik_student()}},
\code{\link[=log_lik_unif]{log_lik_unif()}}
}
\concept{log_lik_dist}
