% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_neg_binom}
\alias{log_lik_neg_binom}
\title{Negative Binomial Log-Likelihood}
\usage{
log_lik_neg_binom(x, lambda, theta)
}
\arguments{
\item{x}{A vector of 0s and 1s.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the shape and rate.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Negative Binomial Log-Likelihood
}
\examples{
log_lik_neg_binom(c(0, 1, 2), 2, 1)
}
\seealso{
Other log_lik_dist: 
\code{\link{log_lik_bern}()},
\code{\link{log_lik_binom}()},
\code{\link{log_lik_gamma_pois}()},
\code{\link{log_lik_lnorm}()},
\code{\link{log_lik_norm}()},
\code{\link{log_lik_pois}()}
}
\concept{log_lik_dist}
