% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_norm}
\alias{log_lik_norm}
\title{Normal Log-Likelihood}
\usage{
log_lik_norm(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Normal Log-Likelihood
}
\examples{
log_lik_norm(c(-2:2))
}
\seealso{
Other log_lik_dist: 
\code{\link{log_lik_bern}()},
\code{\link{log_lik_beta_binom}()},
\code{\link{log_lik_binom}()},
\code{\link{log_lik_gamma_pois_zi}()},
\code{\link{log_lik_gamma_pois}()},
\code{\link{log_lik_gamma}()},
\code{\link{log_lik_lnorm}()},
\code{\link{log_lik_neg_binom}()},
\code{\link{log_lik_norm_skew}()},
\code{\link{log_lik_pois_zi}()},
\code{\link{log_lik_pois}()},
\code{\link{log_lik_student}()}
}
\concept{log_lik_dist}
