% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_pois}
\alias{log_lik_pois}
\title{Poisson Log-Likelihood}
\usage{
log_lik_pois(x, lambda = 1)
}
\arguments{
\item{x}{A non-negative whole numeric vector of values.}

\item{lambda}{A non-negative numeric vector of means.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Poisson Log-Likelihood
}
\examples{
log_lik_pois(c(1,3.5,4), 3)
}
\seealso{
Other log_lik_dist: 
\code{\link{log_lik_bern}()},
\code{\link{log_lik_binom}()},
\code{\link{log_lik_gamma_pois}()},
\code{\link{log_lik_lnorm}()},
\code{\link{log_lik_neg_binom}()},
\code{\link{log_lik_norm}()},
\code{\link{log_lik_pois_zi}()}
}
\concept{log_lik_dist}
