% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_skewlnorm}
\alias{log_lik_skewlnorm}
\title{Skew-Lognormal Log-Likelihood}
\usage{
log_lik_skewlnorm(
  x,
  meanlog = 0,
  sdlog = 1,
  shape = 0,
  tlower = 0,
  tupper = Inf
)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the
log scale.}

\item{shape}{A numeric vector of shape.}

\item{tlower}{A numeric vector of the lower truncation point.}

\item{tupper}{A numeric vector of the upper truncation point.}
}
\value{
An numeric vector of the corresponding log-likelihoods.
}
\description{
Skew-Lognormal Log-Likelihood
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
log_lik_skewlnorm(1:5)
log_lik_skewlnorm(1:5, shape = -2)
log_lik_skewlnorm(1:5, shape = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other log_lik_dist:
\code{\link[=log_lik_bern]{log_lik_bern()}},
\code{\link[=log_lik_beta]{log_lik_beta()}},
\code{\link[=log_lik_beta_binom]{log_lik_beta_binom()}},
\code{\link[=log_lik_binom]{log_lik_binom()}},
\code{\link[=log_lik_exp]{log_lik_exp()}},
\code{\link[=log_lik_gamma]{log_lik_gamma()}},
\code{\link[=log_lik_gamma_pois]{log_lik_gamma_pois()}},
\code{\link[=log_lik_gamma_pois_zi]{log_lik_gamma_pois_zi()}},
\code{\link[=log_lik_lnorm]{log_lik_lnorm()}},
\code{\link[=log_lik_multinom]{log_lik_multinom()}},
\code{\link[=log_lik_neg_binom]{log_lik_neg_binom()}},
\code{\link[=log_lik_norm]{log_lik_norm()}},
\code{\link[=log_lik_pois]{log_lik_pois()}},
\code{\link[=log_lik_pois_zi]{log_lik_pois_zi()}},
\code{\link[=log_lik_skewnorm]{log_lik_skewnorm()}},
\code{\link[=log_lik_student]{log_lik_student()}},
\code{\link[=log_lik_unif]{log_lik_unif()}}
}
\concept{log_lik_dist}
