% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit<-}
\alias{logit<-}
\title{Logistic Transformation}
\usage{
logit(x) <- value
}
\arguments{
\item{x}{An existing object.}

\item{value}{A numeric atomic object of the value to
inverse logistically transform.}
}
\value{
Called for the side effect of updating \code{x}.
}
\description{
Logistic Transformation
}
\details{
A wrapper on \code{stats::plogis(value)}.
}
\examples{
x <- 1
logit(x) <- 0.5
x
}
\seealso{
Other translations: 
\code{\link{fabs}()},
\code{\link{ilogit}()},
\code{\link{invlogit}()},
\code{\link{log<-}()},
\code{\link{logit}()},
\code{\link{odds}()},
\code{\link{phi}()},
\code{\link{pow}()}
}
\concept{translations}
