% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logistic Transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{A numeric atomic object.}
}
\value{
The logistically transformed numeric atomic object.
}
\description{
Logistic transforms a numeric atomic object.
}
\details{
A wrapper on \code{\link[stats:Logistic]{stats::qlogis()}}.
}
\examples{
logit(c(0.25, 0.5, 0.75))
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{exp2}()},
\code{\link{fabs}()},
\code{\link{ilog10}()},
\code{\link{ilog2}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log10<-}()},
\code{\link{log2<-}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{phi}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
