% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{Logistic Transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{A numeric atomic object.}
}
\value{
The logistically transformed numeric atomic object.
}
\description{
Performs the logistic transformation on a numeric atomic object.
}
\details{
A wrapper on \code{stats::plogis()}.
}
\examples{
logit(c(0.25, 0.5, 0.75))
}
\seealso{
Other {translations}: 
\code{\link{ilogit}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{phi}()},
\code{\link{pow}()}
}
\concept{{translations}}
