% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{lower}
\alias{lower}
\title{Lower Credible Limit}
\usage{
lower(x, conf_level = 0.95)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}

\item{conf_level}{A numeric scalar between 0 and 1 specifying the confidence level.}
}
\value{
A number.
}
\description{
Lower Credible Limit
}
\examples{
lower(as.numeric(0:100))
}
\seealso{
Other {MCMC vector to scalar functions}: 
\code{\link{pvalue}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{zscore}()}
}
\concept{{MCMC vector to scalar functions}}
