% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims.default}
\alias{ndims.default}
\title{Number of Dimensions}
\usage{
\method{ndims}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of dimensions.
}
\description{
Gets the number of dimensions of an object as returned by \code{dims()}.
}
\examples{
ndims(character(0))
ndims(1:3)
}
\seealso{
Other {dimensions}: 
\code{\link[universals]{dims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims}()}
}
