% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims.matrix}
\alias{ndims.matrix}
\title{Dimensions of a Matrix}
\usage{
\method{ndims}{matrix}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of dimensions.
}
\description{
Gets the number of dimensions of an object as defined by \code{\link[universals:dims]{dims()}}.
}
\details{
Always 2L.
}
\examples{
ndims(matrix(1))
}
\seealso{
\code{\link[universals:dims]{dims()}}
}
