% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsams.R
\name{nsams.default}
\alias{nsams.default}
\title{Number of Samples}
\usage{
\method{nsams}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of samples.
}
\description{
Gets the number of sample values (simulations * terms)
of an MCMC object.
}
\examples{
nsams.foobar <- function(x, ...) {
  nsims(x, ...) * nterms(x, ...)
}
}
\seealso{
Other MCMC dimensions: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}
}
