% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds}
\alias{odds}
\title{Odds}
\usage{
odds(x)
}
\arguments{
\item{x}{A numeric object (vector, matrix or array) of probabilities.}
}
\value{
A numeric object of the the odds for each probability.
}
\description{
Calculates the odds for probabilities.
}
\examples{
odds(c(0, 0.5, 0.9, 1))
}
\seealso{
Other odds: 
\code{\link{log_odds_ratio}()},
\code{\link{odds_ratio}()}
}
\concept{odds}
