% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par-pattern.R
\name{par_pattern}
\alias{par_pattern}
\title{Parameter Pattern}
\usage{
par_pattern()
}
\value{
A string of the regular expression for a parameter name.
}
\description{
Parameter Pattern
}
\examples{
par_pattern()
}
