% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par-pattern.R
\name{par_pattern}
\alias{par_pattern}
\title{Parameter Pattern}
\usage{
par_pattern(partial = FALSE)
}
\arguments{
\item{partial}{A flag specifying whether the pattern is a partial pattern.
By default FALSE.}
}
\value{
A string of the regular expression for a parameter name.
}
\description{
Parameter Pattern
}
\examples{
par_pattern()
par_pattern(partial = TRUE)
}
