% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}

\item{nas}{A flag specifying whether to also fill missing values.}
}
\description{
Parameter Descriptions
}
