% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}

\item{nas}{A flag specifying whether to also fill missing values.}

\item{conf_level}{A numeric scalar between 0 and 1
specifying the confidence level.}

\item{value}{A scalar of the value to replace values with.}

\item{partial}{A flag specifying whether the pattern is a partial pattern.
By default FALSE.}
}
\description{
Parameter Descriptions
}
\keyword{internal}
