% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi}
\alias{phi}
\title{Phi}
\usage{
phi(x)
}
\arguments{
\item{x}{A numeric atomic object.}
}
\value{
A numeric atomic object.
}
\description{
The standard normal cumulative density function.
}
\details{
A wrapper on \code{\link[stats:Normal]{stats::pnorm()}}.
}
\examples{
phi(0:2)
}
\seealso{
Other translations: 
\code{\link{exp10}()},
\code{\link{exp2}()},
\code{\link{fabs}()},
\code{\link{ilog10}()},
\code{\link{ilog2}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{inv_logit}()},
\code{\link{invlogit}()},
\code{\link{log10<-}()},
\code{\link{log2<-}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{pow}()},
\code{\link{step}()}
}
\concept{translations}
