% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi}
\alias{phi}
\title{Phi}
\usage{
phi(x)
}
\arguments{
\item{x}{A numeric atomic object.}
}
\value{
A numeric atomic object.
}
\description{
The standard normal cumulative density function.
}
\details{
A wrapper on \code{\link[stats:Normal]{stats::pnorm()}}.
}
\examples{
phi(0:2)
}
\seealso{
Other translations: 
\code{\link{fabs}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{invlogit}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{pow}()}
}
\concept{translations}
