% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Power}
\usage{
pow(x, n)
}
\arguments{
\item{x}{A numeric atomic object of the base.}

\item{n}{A numeric atomic object of the exponent.}
}
\value{
A numeric atomic object of x raised to n.
}
\description{
R equivalent to the power function.
}
\details{
Wrapper on \code{x^n}.
}
\examples{
pow(10, 2)
}
\seealso{
Other translations: 
\code{\link{fabs}()},
\code{\link{ilogit}()},
\code{\link{ilog}()},
\code{\link{invlogit}()},
\code{\link{log<-}()},
\code{\link{logit<-}()},
\code{\link{logit}()},
\code{\link{phi}()}
}
\concept{translations}
