% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_bern}
\alias{prob_bern}
\title{Bernoulli Cumulative Distribution Function}
\usage{
prob_bern(x, prob = 0.5)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of
success.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Bernoulli Cumulative Distribution Function
}
\examples{
prob_bern(c(TRUE, FALSE), 0.7)
}
\seealso{
Other prob_dist: 
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
