% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_gamma}
\alias{prob_gamma}
\title{Gamma Cumulative Distribution Function}
\usage{
prob_gamma(x, shape = 1, rate = 1)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{shape}{A non-negative numeric vector of shape.}

\item{rate}{A non-negative numeric vector of rate.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Gamma Cumulative Distribution Function
}
\examples{
prob_gamma(c(0, 1, 2), 1, 2)
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_skewnorm}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
