% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_skewlnorm}
\alias{prob_skewlnorm}
\title{Skew-Lognormal Cumulative Distribution Function}
\usage{
prob_skewlnorm(
  x,
  location_log = 0,
  scale_log = 1,
  shape_log = 0,
  ...,
  meanlog,
  sdlog,
  shape
)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{location_log}{A numeric vector of location parameters of \code{log(x)}.}

\item{scale_log}{A non-negative numeric vector of scale parameters of \code{log(x)}.}

\item{shape_log}{A numeric vector of shape parameters of \code{log(x)}.
Negative values result in leftward skew, while positive values result in rightward skew.}

\item{...}{Unused.}

\item{meanlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of location parameters of \code{log(x)}.
Described as "a numeric vector of the means on the log scale" prior to v.
0.10.1.
Will be removed in a future version.}

\item{sdlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of scale parameters of
\code{log(x)}.
Described as "a non-negative numeric vector of the standard deviations on the
log scale" prior to v. 0.10.1.
Will be removed in a future version.}

\item{shape}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of shape parameters of \code{log(x)}.
Will be removed in a future version.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Skew-Lognormal Cumulative Distribution Function
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
prob_skewlnorm(1:5)
prob_skewlnorm(1:5, shape_log = -2)
prob_skewlnorm(1:5, shape_log = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other prob_dist:
\code{\link[=prob_bern]{prob_bern()}},
\code{\link[=prob_beta]{prob_beta()}},
\code{\link[=prob_beta_binom]{prob_beta_binom()}},
\code{\link[=prob_binom]{prob_binom()}},
\code{\link[=prob_exp]{prob_exp()}},
\code{\link[=prob_gamma]{prob_gamma()}},
\code{\link[=prob_gamma_pois]{prob_gamma_pois()}},
\code{\link[=prob_gamma_pois_zi]{prob_gamma_pois_zi()}},
\code{\link[=prob_lnorm]{prob_lnorm()}},
\code{\link[=prob_neg_binom]{prob_neg_binom()}},
\code{\link[=prob_norm]{prob_norm()}},
\code{\link[=prob_pois]{prob_pois()}},
\code{\link[=prob_pois_zi]{prob_pois_zi()}},
\code{\link[=prob_skewnorm]{prob_skewnorm()}},
\code{\link[=prob_student]{prob_student()}},
\code{\link[=prob_unif]{prob_unif()}}
}
\concept{prob_dist}
