% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob_skewnorm}
\alias{prob_skewnorm}
\title{Skew Normal Cumulative Distribution Function}
\usage{
prob_skewnorm(x, mean = 0, sd = 1, shape = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{shape}{A numeric vector of shape.}
}
\value{
An numeric vector of the corresponding probabilities.
}
\description{
Skew Normal Cumulative Distribution Function
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
prob_skewnorm(c(-2:2))
prob_skewnorm(c(-2:2), shape = -2)
prob_skewnorm(c(-2:2), shape = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other prob_dist: 
\code{\link{prob_bern}()},
\code{\link{prob_beta_binom}()},
\code{\link{prob_binom}()},
\code{\link{prob_gamma}()},
\code{\link{prob_gamma_pois}()},
\code{\link{prob_gamma_pois_zi}()},
\code{\link{prob_lnorm}()},
\code{\link{prob_neg_binom}()},
\code{\link{prob_norm}()},
\code{\link{prob_pois}()},
\code{\link{prob_pois_zi}()},
\code{\link{prob_student}()}
}
\concept{prob_dist}
