% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{Bayesian P-Value}
\usage{
pvalue(x)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}
}
\value{
A number between 0 and 1.
}
\description{
Bayesian P-Value
}
\examples{
pvalue(as.numeric(0:100))
}
\seealso{
\code{\link{svalue}}
}
