% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_bern}
\alias{quant_bern}
\title{Bernoulli Quantile Function}
\usage{
quant_bern(x, prob = 0.5)
}
\arguments{
\item{x}{A vector of probabilities.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of
success.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Bernoulli Quantile Function
}
\examples{
quant_bern(c(TRUE, FALSE), 0.7)
}
\seealso{
Other quant_dist: 
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
