% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_gamma_pois}
\alias{quant_gamma_pois}
\title{Gamma-Poisson Quantile Function}
\usage{
quant_gamma_pois(x, lambda = 1, theta = 0)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion for the mixture
models (student, gamma-Poisson and beta-binomial).}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Gamma-Poisson Quantile Function
}
\examples{
quant_gamma_pois(c(0.1, 0.4, 0.6), 1, 1)
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
