% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_lnorm}
\alias{quant_lnorm}
\title{Log-Normal Quantile Function}
\usage{
quant_lnorm(x, meanlog = 0, sdlog = 1)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the
log scale.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Log-Normal Quantile Function
}
\examples{
quant_lnorm(c(0.1, 0.4, 0.6), 0, 2)
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_skewnorm}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
