% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_skewlnorm}
\alias{quant_skewlnorm}
\title{Skew-Lognormal Quantile Function}
\usage{
quant_skewlnorm(
  x,
  location_log = 0,
  scale_log = 1,
  shape_log = 0,
  ...,
  meanlog,
  sdlog,
  shape
)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{location_log}{A numeric vector of location parameters of \code{log(x)}.}

\item{scale_log}{A non-negative numeric vector of scale parameters of \code{log(x)}.}

\item{shape_log}{A numeric vector of shape parameters of \code{log(x)}.
Negative values result in leftward skew, while positive values result in rightward skew.}

\item{...}{Unused.}

\item{meanlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of location parameters of \code{log(x)}.
Described as "a numeric vector of the means on the log scale" prior to v.
0.10.1.
Will be removed in a future version.}

\item{sdlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of scale parameters of
\code{log(x)}.
Described as "a non-negative numeric vector of the standard deviations on the
log scale" prior to v. 0.10.1.
Will be removed in a future version.}

\item{shape}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of shape parameters of \code{log(x)}.
Will be removed in a future version.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Skew-Lognormal Quantile Function
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
quant_skewlnorm(c(0.1, 0.4, 0.6))
quant_skewlnorm(c(0.1, 0.4, 0.6), shape_log = -2)
quant_skewlnorm(c(0.1, 0.4, 0.6), shape_log = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other quant_dist:
\code{\link[=quant_bern]{quant_bern()}},
\code{\link[=quant_beta]{quant_beta()}},
\code{\link[=quant_binom]{quant_binom()}},
\code{\link[=quant_exp]{quant_exp()}},
\code{\link[=quant_gamma]{quant_gamma()}},
\code{\link[=quant_gamma_pois]{quant_gamma_pois()}},
\code{\link[=quant_gamma_pois_zi]{quant_gamma_pois_zi()}},
\code{\link[=quant_lnorm]{quant_lnorm()}},
\code{\link[=quant_neg_binom]{quant_neg_binom()}},
\code{\link[=quant_norm]{quant_norm()}},
\code{\link[=quant_pois]{quant_pois()}},
\code{\link[=quant_pois_zi]{quant_pois_zi()}},
\code{\link[=quant_skewnorm]{quant_skewnorm()}},
\code{\link[=quant_student]{quant_student()}},
\code{\link[=quant_unif]{quant_unif()}}
}
\concept{quant_dist}
