% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant_skewnorm}
\alias{quant_skewnorm}
\title{Skew Normal Quantile Function}
\usage{
quant_skewnorm(x, mean = 0, sd = 1, shape = 0)
}
\arguments{
\item{x}{A numeric vector of probabilities.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{shape}{A numeric vector of shape.}
}
\value{
An numeric vector of the corresponding quantiles.
}
\description{
Skew Normal Quantile Function
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
quant_skewnorm(c(-2:2))
quant_skewnorm(c(-2:2), shape = -2)
quant_skewnorm(c(-2:2), shape = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other quant_dist: 
\code{\link{quant_bern}()},
\code{\link{quant_beta_binom}()},
\code{\link{quant_binom}()},
\code{\link{quant_gamma}()},
\code{\link{quant_gamma_pois}()},
\code{\link{quant_gamma_pois_zi}()},
\code{\link{quant_lnorm}()},
\code{\link{quant_neg_binom}()},
\code{\link{quant_norm}()},
\code{\link{quant_pois}()},
\code{\link{quant_pois_zi}()},
\code{\link{quant_student}()}
}
\concept{quant_dist}
