% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_gamma_pois_zi}
\alias{ran_gamma_pois_zi}
\title{Zero-Inflated Gamma-Poisson Random Samples}
\usage{
ran_gamma_pois_zi(n = 1, lambda = 1, theta = 0, prob = 0)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to generate.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the dispersion parameters for the gamma mixing distribution. The variance for the resultant gamma-Poisson distribution is lambda + lambda^2 * theta.}

\item{prob}{A numeric vector of values between 0 and 1 of the probability of success.}
}
\value{
A numeric vector of the random samples.
}
\description{
Zero-Inflated Gamma-Poisson Random Samples
}
\examples{
ran_gamma_pois_zi(10, lambda = 3, theta = 1, prob = 0.5)
}
\seealso{
Other ran_dist: 
\code{\link{ran_bern}()},
\code{\link{ran_binom}()},
\code{\link{ran_gamma_pois}()},
\code{\link{ran_gamma}()},
\code{\link{ran_lnorm}()},
\code{\link{ran_neg_binom}()},
\code{\link{ran_norm}()},
\code{\link{ran_pois_zi}()},
\code{\link{ran_pois}()}
}
\concept{ran_dist}
