% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_lnorm}
\alias{ran_lnorm}
\title{Log-Normal Random Samples}
\usage{
ran_lnorm(n = 1, meanlog = 0, sdlog = 1)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to generate.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the log scale.}
}
\value{
A numeric vector of the random samples.
}
\description{
Log-Normal Random Samples
}
\examples{
ran_lnorm(10)
}
\seealso{
Other ran_dist: 
\code{\link{ran_bern}()},
\code{\link{ran_binom}()},
\code{\link{ran_gamma_pois_zi}()},
\code{\link{ran_gamma_pois}()},
\code{\link{ran_gamma}()},
\code{\link{ran_neg_binom}()},
\code{\link{ran_norm}()},
\code{\link{ran_pois_zi}()},
\code{\link{ran_pois}()}
}
\concept{ran_dist}
