% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_neg_binom}
\alias{ran_neg_binom}
\title{Gamma-Poisson Random Samples}
\usage{
ran_neg_binom(n = 1, lambda = 1, theta = 0)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to generate.}

\item{lambda}{A non-negative numeric vector of means.}

\item{theta}{A non-negative numeric vector of the shape and rate.}
}
\value{
An numeric vector of the random samples.
}
\description{
Gamma-Poisson Random Samples
}
\examples{
ran_neg_binom(10, 1, 1)
}
\seealso{
Other ran_dist: 
\code{\link{ran_bern}()},
\code{\link{ran_binom}()},
\code{\link{ran_gamma_pois}()},
\code{\link{ran_gamma}()},
\code{\link{ran_lnorm}()},
\code{\link{ran_norm}()},
\code{\link{ran_pois}()}
}
\concept{ran_dist}
