% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_norm}
\alias{ran_norm}
\title{Normal Random Samples}
\usage{
ran_norm(n = 1, mean = 0, sd = 1)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to generate.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}
}
\value{
A numeric vector of the random samples.
}
\description{
Normal Random Samples
}
\examples{
ran_norm(10)
}
\seealso{
Other ran_dist: 
\code{\link{ran_bern}()},
\code{\link{ran_beta_binom}()},
\code{\link{ran_binom}()},
\code{\link{ran_gamma_pois_zi}()},
\code{\link{ran_gamma_pois}()},
\code{\link{ran_gamma}()},
\code{\link{ran_lnorm}()},
\code{\link{ran_neg_binom}()},
\code{\link{ran_norm_skew}()},
\code{\link{ran_pois_zi}()},
\code{\link{ran_pois}()},
\code{\link{ran_student}()}
}
\concept{ran_dist}
