% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_skewlnorm}
\alias{ran_skewlnorm}
\title{Skew-Lognormal Random Samples}
\usage{
ran_skewlnorm(
  n = 1,
  location_log = 0,
  scale_log = 1,
  shape_log = 0,
  ...,
  meanlog,
  sdlog,
  shape
)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to
generate.}

\item{location_log}{A numeric vector of location parameters of \code{log(x)}.}

\item{scale_log}{A non-negative numeric vector of scale parameters of \code{log(x)}.}

\item{shape_log}{A numeric vector of shape parameters of \code{log(x)}.
Negative values result in leftward skew, while positive values result in rightward skew.}

\item{...}{Unused.}

\item{meanlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of location parameters of \code{log(x)}.
Described as "a numeric vector of the means on the log scale" prior to v.
0.10.1.
Will be removed in a future version.}

\item{sdlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of scale parameters of
\code{log(x)}.
Described as "a non-negative numeric vector of the standard deviations on the
log scale" prior to v. 0.10.1.
Will be removed in a future version.}

\item{shape}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of shape parameters of \code{log(x)}.
Will be removed in a future version.}
}
\value{
A numeric vector of the random samples.
}
\description{
Skew-Lognormal Random Samples
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
ran_skewlnorm(10, shape_log = -1)
ran_skewlnorm(10, shape_log = 0)
ran_skewlnorm(10, shape_log = 1)
\dontshow{\}) # examplesIf}
}
\seealso{
Other ran_dist:
\code{\link[=ran_bern]{ran_bern()}},
\code{\link[=ran_beta_binom]{ran_beta_binom()}},
\code{\link[=ran_binom]{ran_binom()}},
\code{\link[=ran_gamma]{ran_gamma()}},
\code{\link[=ran_gamma_pois]{ran_gamma_pois()}},
\code{\link[=ran_gamma_pois_zi]{ran_gamma_pois_zi()}},
\code{\link[=ran_lnorm]{ran_lnorm()}},
\code{\link[=ran_multinom]{ran_multinom()}},
\code{\link[=ran_neg_binom]{ran_neg_binom()}},
\code{\link[=ran_norm]{ran_norm()}},
\code{\link[=ran_pois]{ran_pois()}},
\code{\link[=ran_pois_zi]{ran_pois_zi()}},
\code{\link[=ran_skewnorm]{ran_skewnorm()}},
\code{\link[=ran_student]{ran_student()}}
}
\concept{ran_dist}
