% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_skewlnorm}
\alias{ran_skewlnorm}
\title{Skew-Lognormal Random Samples}
\usage{
ran_skewlnorm(n = 1, meanlog = 0, sdlog = 1, shape = 0)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to
generate.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the
log scale.}

\item{shape}{A numeric vector of shape.}
}
\value{
A numeric vector of the random samples.
}
\description{
Skew-Lognormal Random Samples
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
ran_skewlnorm(10, shape = -1)
ran_skewlnorm(10, shape = 0)
ran_skewlnorm(10, shape = 1)
\dontshow{\}) # examplesIf}
}
\seealso{
Other ran_dist:
\code{\link[=ran_bern]{ran_bern()}},
\code{\link[=ran_beta_binom]{ran_beta_binom()}},
\code{\link[=ran_binom]{ran_binom()}},
\code{\link[=ran_gamma]{ran_gamma()}},
\code{\link[=ran_gamma_pois]{ran_gamma_pois()}},
\code{\link[=ran_gamma_pois_zi]{ran_gamma_pois_zi()}},
\code{\link[=ran_lnorm]{ran_lnorm()}},
\code{\link[=ran_neg_binom]{ran_neg_binom()}},
\code{\link[=ran_norm]{ran_norm()}},
\code{\link[=ran_pois]{ran_pois()}},
\code{\link[=ran_pois_zi]{ran_pois_zi()}},
\code{\link[=ran_skewnorm]{ran_skewnorm()}},
\code{\link[=ran_student]{ran_student()}}
}
\concept{ran_dist}
