% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran.R
\name{ran_skewnorm}
\alias{ran_skewnorm}
\title{Skew Normal Random Samples}
\usage{
ran_skewnorm(n = 1, mean = 0, sd = 1, shape = 0)
}
\arguments{
\item{n}{A non-negative whole number of the number of random samples to
generate.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{shape}{A numeric vector of shape.}
}
\value{
A numeric vector of the random samples.
}
\description{
Skew Normal Random Samples
}
\examples{
\dontshow{if (rlang::is_installed("sn")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ran_skewnorm(10, shape = -1)
ran_skewnorm(10, shape = 0)
ran_skewnorm(10, shape = 1)
\dontshow{\}) # examplesIf}
}
\seealso{
Other ran_dist: 
\code{\link{ran_bern}()},
\code{\link{ran_beta_binom}()},
\code{\link{ran_binom}()},
\code{\link{ran_gamma}()},
\code{\link{ran_gamma_pois}()},
\code{\link{ran_gamma_pois_zi}()},
\code{\link{ran_lnorm}()},
\code{\link{ran_neg_binom}()},
\code{\link{ran_norm}()},
\code{\link{ran_pois}()},
\code{\link{ran_pois_zi}()},
\code{\link{ran_student}()}
}
\concept{ran_dist}
