% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R, R/fill-all.R, R/fill-na.R, R/ndims.R,
%   R/npars.R, R/npdims.R, R/nsams.R, R/nsims.R, R/numericise.R, R/set-pars.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dims}
\alias{fill_all}
\alias{fill_na}
\alias{ndims}
\alias{npars}
\alias{npdims}
\alias{nsams}
\alias{nsims}
\alias{numericise}
\alias{numericize}
\alias{set_pars}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{universals}{\code{\link[universals]{dims}}, \code{\link[universals]{fill_all}}, \code{\link[universals]{fill_na}}, \code{\link[universals]{ndims}}, \code{\link[universals]{npars}}, \code{\link[universals]{npdims}}, \code{\link[universals]{nsams}}, \code{\link[universals]{nsims}}, \code{\link[universals]{numericise}}, \code{\link[universals]{numericize}}, \code{\link[universals]{set_pars}}}
}}

