% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{res_norm}
\alias{res_norm}
\title{Normal Residuals}
\usage{
res_norm(x, mean = 0, sd = 1, type = "dev", resample = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{type}{A string of the residual type ('raw' or 'dev').}

\item{resample}{A flag specifying whether to resample.}
}
\value{
An numeric vector of the corresponding residuals.
}
\description{
Normal Residuals
}
\examples{
dev_norm(c(-2:2))
}
\seealso{
Other res_dist: 
\code{\link{res_bern}()},
\code{\link{res_binom}()},
\code{\link{res_gamma_pois}()},
\code{\link{res_lnorm}()},
\code{\link{res_pois}()}
}
\concept{res_dist}
