% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{res_skewlnorm}
\alias{res_skewlnorm}
\title{Skew-Lognormal Residuals}
\usage{
res_skewlnorm(
  x,
  location_log = 0,
  scale_log = 1,
  shape_log = 0,
  type = "dev",
  simulate = FALSE,
  ...,
  meanlog,
  sdlog,
  shape
)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{location_log}{A numeric vector of location parameters of \code{log(x)}.}

\item{scale_log}{A non-negative numeric vector of scale parameters of \code{log(x)}.}

\item{shape_log}{A numeric vector of shape parameters of \code{log(x)}.
Negative values result in leftward skew, while positive values result in rightward skew.}

\item{type}{A string of the residual type. 'raw' for raw residuals 'dev' for
deviance residuals and 'data' for the data.}

\item{simulate}{A flag specifying whether to simulate residuals.}

\item{...}{Unused.}

\item{meanlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of location parameters of \code{log(x)}.
Described as "a numeric vector of the means on the log scale" prior to v.
0.10.1.
Will be removed in a future version.}

\item{sdlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A non-negative numeric vector of scale parameters of
\code{log(x)}.
Described as "a non-negative numeric vector of the standard deviations on the
log scale" prior to v. 0.10.1.
Will be removed in a future version.}

\item{shape}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A numeric vector of shape parameters of \code{log(x)}.
Will be removed in a future version.}
}
\value{
An numeric vector of the corresponding residuals.
}
\description{
Skew-Lognormal Residuals
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
res_skewlnorm(exp(-2:2))
\dontshow{\}) # examplesIf}
}
\seealso{
Other res_dist:
\code{\link[=res_bern]{res_bern()}},
\code{\link[=res_beta_binom]{res_beta_binom()}},
\code{\link[=res_binom]{res_binom()}},
\code{\link[=res_gamma]{res_gamma()}},
\code{\link[=res_gamma_pois]{res_gamma_pois()}},
\code{\link[=res_gamma_pois_zi]{res_gamma_pois_zi()}},
\code{\link[=res_lnorm]{res_lnorm()}},
\code{\link[=res_multinom]{res_multinom()}},
\code{\link[=res_neg_binom]{res_neg_binom()}},
\code{\link[=res_norm]{res_norm()}},
\code{\link[=res_pois]{res_pois()}},
\code{\link[=res_pois_zi]{res_pois_zi()}},
\code{\link[=res_skewnorm]{res_skewnorm()}},
\code{\link[=res_student]{res_student()}}
}
\concept{res_dist}
