% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_beta}
\alias{sens_beta}
\title{Adjust Beta Distribution Parameters for Sensitivity Analyses}
\usage{
sens_beta(alpha, beta, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Beta Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_beta(10, 10, 2)
sens_beta(10, 10, 0.8)
}
