% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_exp}
\alias{sens_exp}
\title{Adjust Exponential Distribution Parameters for Sensitivity Analyses}
\usage{
sens_exp(rate, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Exponential Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_exp(10, 2)
sens_exp(10, 0.8)
}
