% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_gamma_pois}
\alias{sens_gamma_pois}
\title{Adjust Gamma-Poisson Distribution Parameters for Sensitivity Analyses}
\usage{
sens_gamma_pois(lambda, theta, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Gamma-Poisson Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_gamma_pois(10, 0.1, 2)
}
