% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_norm}
\alias{sens_norm}
\title{Adjust Normal Distribution Parameters for Sensitivity Analyses}
\usage{
sens_norm(mean, sd, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Normal Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_norm(10, 3, 2)
sens_norm(10, 3, 0.8)
}
