% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_skewnorm}
\alias{sens_skewnorm}
\title{Adjust Skew Normal Distribution Parameters for Sensitivity Analyses}
\usage{
sens_skewnorm(mean, sd, shape, sd_mult = 2)
}
\arguments{
\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{shape}{A numeric vector of shape.}

\item{sd_mult}{A non-negative multiplier on the standard deviation of the
distribution.}
}
\value{
A named list of the adjusted distribution's parameters.
}
\description{
Expands (\code{sd_mult > 1}) or reduces (\code{sd_mult < 1}) the standard deviation
of the Skew Normal distribution without changing the mean.
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
sens_skewnorm(10, 3, -1, 2)
sens_skewnorm(10, 3, 3, 0.8)
\dontshow{\}) # examplesIf}
}
\seealso{
Other sens_dist:
\code{\link[=sens_beta]{sens_beta()}},
\code{\link[=sens_exp]{sens_exp()}},
\code{\link[=sens_gamma]{sens_gamma()}},
\code{\link[=sens_gamma_pois]{sens_gamma_pois()}},
\code{\link[=sens_gamma_pois_zi]{sens_gamma_pois_zi()}},
\code{\link[=sens_lnorm]{sens_lnorm()}},
\code{\link[=sens_neg_binom]{sens_neg_binom()}},
\code{\link[=sens_norm]{sens_norm()}},
\code{\link[=sens_pois]{sens_pois()}},
\code{\link[=sens_skewlnorm]{sens_skewlnorm()}},
\code{\link[=sens_student]{sens_student()}}
}
\concept{sens_dist}
