% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_skewnorm}
\alias{sens_skewnorm}
\title{Adjust Skew Normal Distribution Parameters for Sensitivity Analyses}
\usage{
sens_skewnorm(mean, sd, shape, sd_mult = 2)
}
\arguments{
\item{sd_mult}{}
}
\value{
A named vector of the adjusted distribution's parameters.
}
\description{
Adjust Skew Normal Distribution Parameters for Sensitivity Analyses
}
\examples{
sens_skewnorm(10, 3, -1, 2)
sens_skewnorm(10, 3, 3, 0.8)
}
