% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_skewnorm}
\alias{sens_skewnorm}
\title{Adjust Skew Normal Distribution Parameters for Sensitivity Analyses}
\usage{
sens_skewnorm(mean, sd, shape, sd_mult = 2)
}
\arguments{
\item{mean}{A numeric vector of the means.}

\item{sd}{A non-negative numeric vector of the standard deviations.}

\item{shape}{A non-negative numeric vector of shape.}

\item{sd_mult}{A non-negative multiplier on the standard deviation of the
distribution.}
}
\value{
A named list of the adjusted distribution's parameters.
}
\description{
Expands (\code{sd_mult > 1}) or reduces (\code{sd_mult < 1}) the standard deviation
of the Skew Normal distribution without changing the mean.
}
\examples{
\dontshow{if (rlang::is_installed("sn")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sens_skewnorm(10, 3, -1, 2)
sens_skewnorm(10, 3, 3, 0.8)
\dontshow{\}) # examplesIf}
}
\seealso{
Other sens_dist: 
\code{\link{sens_beta}()},
\code{\link{sens_exp}()},
\code{\link{sens_gamma}()},
\code{\link{sens_gamma_pois}()},
\code{\link{sens_gamma_pois_zi}()},
\code{\link{sens_lnorm}()},
\code{\link{sens_neg_binom}()},
\code{\link{sens_norm}()},
\code{\link{sens_pois}()},
\code{\link{sens_student}()}
}
\concept{sens_dist}
