% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewlnorm.R
\name{dskewlnorm}
\alias{dskewlnorm}
\alias{pskewlnorm}
\alias{qskewlnorm}
\alias{rskewlnorm}
\title{Skew-Lognormal Distribution}
\usage{
dskewlnorm(x, meanlog = 0, sdlog = 1, shape = 0, log = FALSE)

pskewlnorm(q, meanlog = 0, sdlog = 1, shape = 0)

qskewlnorm(p, meanlog = 0, sdlog = 1, shape = 0)

rskewlnorm(n = 1, meanlog = 0, sdlog = 1, shape = 0)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{meanlog}{A numeric vector of the means on the log scale.}

\item{sdlog}{A non-negative numeric vector of the standard deviations on the
log scale.}

\item{shape}{A numeric vector of values.}

\item{log}{A flag specifying whether to return the log-transformed value.}

\item{q}{A vector of quantiles.}

\item{p}{A vector of probabilities.}

\item{n}{A non-negative whole number of the number of random samples to
generate.}
}
\value{
\code{dskewlnorm} gives the density, \code{pskewlnorm} gives the distribution function, \code{qskewlnorm} gives the quantile function, and \code{rskewlnorm} generates random deviates.
\code{pskewlnorm} and \code{qskewlnorm} use the lower tail probability.
}
\description{
The skew-lognormal distribution of a value \code{x} whose natural logarithm
follows a \link[=dskewnorm]{Skew-Normal} distribution with location \code{meanlog},
scale \code{sdlog} and \code{shape}.
It reduces to the Log-Normal distribution when \code{shape = 0}.
}
\examples{
\dontshow{if (rlang::is_installed("sn")) withAutoprint(\{ # examplesIf}
dskewlnorm(x = 1:5, meanlog = 0, sdlog = 1, shape = 0.1)
dskewlnorm(x = 1:5, meanlog = 0, sdlog = 1, shape = -1)
qskewlnorm(p = c(0.1, 0.4), meanlog = 0, sdlog = 1, shape = 0.1)
qskewlnorm(p = c(0.1, 0.4), meanlog = 0, sdlog = 1, shape = -1)
pskewlnorm(q = 1:5, meanlog = 0, sdlog = 1, shape = 0.1)
pskewlnorm(q = 1:5, meanlog = 0, sdlog = 1, shape = -1)
rskewlnorm(n = 3, meanlog = 0, sdlog = 1, shape = 0.1)
rskewlnorm(n = 3, meanlog = 0, sdlog = 1, shape = -1)
\dontshow{\}) # examplesIf}
}
\concept{skewlnorm}
