% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{A numeric object of MCMC values.}
}
\value{
A number.
}
\description{
Skewness
}
\examples{
skewness(1:10)
}
\seealso{
Other summary: 
\code{\link{kurtosis}()},
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
