% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svalue.R
\name{svalue}
\alias{svalue}
\title{Surprisal Value}
\usage{
svalue(x, side = NULL, threshold = 0, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric object of MCMC values.}

\item{side}{A character indicating whether to calculate s-values using
p-values for the left tail (\code{"left"}), right tail (\code{"right"}), or both tails
(\code{"both"}).
Defaults to \code{NULL}, which is treated as \code{"both"} but returns a warning.
The warning will be removed in future package versions.}

\item{threshold}{A number of the threshold value.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A non-negative number.
}
\description{
The surprisal value (Greenland 2019) is the \link{pvalue} expressed in
terms of how many consecutive heads would have to be thrown on a fair coin
in a single attempt to achieve the same probability: \eqn{-\log_2(p)}, where
\eqn{p} is the p-value of interest.
}
\examples{
svalue(rnorm(1e4, mean = 1), side = "left")
svalue(rnorm(1e4, mean = 1), side = "right")
}
\references{
Greenland, S. 2019. Valid P-Values Behave Exactly as They Should:
Some Misleading Criticisms of P-Values and Their Resolution With S -Values.
The American Statistician 73(sup1): 106–114.
\doi{10.1080/00031305.2018.1529625}.
}
\seealso{
Other summary:
\code{\link{kurtosis}()},
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{pzeros}()},
\code{\link{skewness}()},
\code{\link{upper}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{xtr_median}()},
\code{\link{xtr_sd}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
