% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svalue.R
\name{svalue}
\alias{svalue}
\title{Bayesian Surprisal Value}
\usage{
svalue(x)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}
}
\value{
A non-negative number.
}
\description{
Bayesian Surprisal Value
}
\examples{
svalue(as.numeric(0:100))
}
\seealso{
Other summary: 
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{upper}()},
\code{\link{zscore}()}
}
\concept{summary}
