% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.R
\name{upper}
\alias{upper}
\title{Upper Credible Limit}
\usage{
upper(x, conf_level = 0.95)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}

\item{conf_level}{A numeric scalar between 0 and 1
specifying the confidence level.}
}
\value{
A number.
}
\description{
Calculates the upper credible limit for a
vector of MCMC values.
}
\details{
By default it returns the 95\% credible limit which
corresponds to the 97.5\% quantile.
}
\examples{
upper(as.numeric(0:100))
}
\seealso{
Other {MCMC vector to scalar functions}: 
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{svalue}()},
\code{\link{zscore}()}
}
\concept{{MCMC vector to scalar functions}}
