% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_eti.R
\name{xtr_ci_eti}
\alias{xtr_ci_eti}
\title{Equal-Tailed Interval}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{level}{A number between 0 and 1 (exclusive) specifying the probability
coverage of the ETI.}

\item{...}{Currently unused.}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of the \code{lower} and \code{upper} limits for the credible interval.
Note that the interval is not guaranteed to be one-sided or two-sided.
}
\description{
Calculates Bayesian credible intervals using the equal-tailed interval (ETI),
i.e., the CI such that the left and right tails outside the CI have the same
coverage.
}
\examples{
xtr_ci_eti(rnorm(1e4))
}
\seealso{
\code{\link[=xtr_ci]{xtr_ci()}} and \code{\link[=xtr_ci_hdi]{xtr_ci_hdi()}}
}
