% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_hdi.R
\name{xtr_ci_hdi}
\alias{xtr_ci_hdi}
\title{Highest Density Interval}
\usage{
xtr_ci_hdi(x, level = 0.95, ..., na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{level}{A number > 0 and <= 1 specifying the probability coverage of the
interval.}

\item{...}{Currently unused.}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A \link{data.frame} of the \code{lower} and \code{upper} limits for the credible
interval.
Note that the interval is not guaranteed to be one-sided or two-sided.
Returns integer limits if the input data are integers and double otherwise.
}
\description{
Calculates Bayesian credible intervals using the highest density
interval (HDI), i.e., the narrowest CI with the specified minimum coverage.
}
\examples{
xtr_ci_hdi(1:10, level = 0.1) # only 10\% of values inside
xtr_ci_hdi(1:10, level = 0.2) # only 20\% of values inside
xtr_ci_hdi(1:10, level = 0.2 + 0.01) # at least 20.1\% of values inside
xtr_ci_hdi(1:100) # inclusive interval [3, 98] with 95\% of values inside
}
\seealso{
\code{\link[=xtr_ci]{xtr_ci()}} and \code{\link[=xtr_ci_eti]{xtr_ci_eti()}}
}
