% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{xtr_median}
\alias{xtr_median}
\title{Median}
\usage{
xtr_median(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric object of MCMC values.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number.
}
\description{
Median
}
\examples{
xtr_mean(1:10)
}
\seealso{
Other summary: 
\code{\link{kurtosis}()},
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{skewness}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
