% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{xtr_rope}
\alias{xtr_rope}
\title{Region of Practical Equivalence}
\usage{
xtr_rope(
  x,
  threshold = 0,
  interval = c(-0.1, 0.1),
  ...,
  level = 1,
  type = "HDI",
  na_rm = FALSE
)
}
\arguments{
\item{x}{A numeric vector of MCMC samples.}

\item{threshold}{A number specifying the center of the ROPE.}

\item{interval}{A numeric vector of length 2 to be added to \code{threshold} to
calculate the ROPE.
Generally, \code{threshold[1] == - threshold[2]} and \code{threshold[1] < 0} should
both be true.}

\item{...}{Currently unused.}

\item{level}{A number > 0 and <= 1 specifying the probability coverage of
the HDI to use.
The default of 1 uses the full posterior.}

\item{type}{A string indicating which type of CI to return.
Currently allows Highest Density Intervals (\code{"HDI"}; default) and
Equal-Tailed Intervals (\code{"ETI"}).}

\item{na_rm}{A flag indicating whether to remove missing values.}
}
\value{
A number indicating the estimated proportion of the posterior (or
credible interval) within the region of practical equivalence.
}
\description{
Calculates the proportion of the posterior (or credible interval if
\code{level < 1}) that falls within the region of practical equivalence (ROPE)
to a \code{threshold}.
The ROPE is calculated as \code{threshold + interval}.
Note that the default is not appropriate for all models, since the \code{interval}
is sensitive to unit conversions.
}
\examples{
xtr_rope(c(-Inf, -1, -0.1, 0.1, 1, Inf))
xtr_rope(rnorm(1e4))
}
\seealso{
\link{xtr_ci}
}
