% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{xtr_sd}
\alias{xtr_sd}
\title{Standard Deviation}
\usage{
xtr_sd(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric object of MCMC values.}

\item{na_rm}{A flag specifying whether to remove missing values.}
}
\value{
A number.
}
\description{
Standard Deviation
}
\examples{
xtr_sd(1:10)
}
\seealso{
Other summary: 
\code{\link{kurtosis}()},
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{skewness}()},
\code{\link{svalue}()},
\code{\link{upper}()},
\code{\link{variance}()},
\code{\link{xtr_mean}()},
\code{\link{xtr_median}()},
\code{\link{zeros}()},
\code{\link{zscore}()}
}
\concept{summary}
