% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Z-Score}
\usage{
zscore(x)
}
\arguments{
\item{x}{A numeric object of MCMC values.}
}
\value{
A number.
}
\description{
The Bayesian z-score is here defined as the number of standard deviations
from the mean estimate to zero.
}
\examples{
zscore(as.numeric(0:100))
}
\seealso{
Other summary: 
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{svalue}()},
\code{\link{upper}()}
}
\concept{summary}
