% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Z-Score}
\usage{
zscore(x)
}
\arguments{
\item{x}{A numeric vector of MCMC values.}
}
\value{
A number.
}
\description{
Z-Score
}
\examples{
zscore(as.numeric(0:100))
}
\seealso{
Other {MCMC vector to scalar functions}: 
\code{\link{lower}()},
\code{\link{pvalue}()},
\code{\link{svalue}()},
\code{\link{upper}()}
}
\concept{{MCMC vector to scalar functions}}
